Working with the FMI model description

The FMI model description provides all human readable information on the model. The following fuctions can be used to obtain all information provided by the model descripton, wich in turn can be extrated from the fmu.

Loading/Parsing

FMIImport.fmi2LoadModelDescription — Function

fmi2LoadModelDescription(pathToModelDescription::String)

Extract the FMU variables and meta data from the ModelDescription

Arguments

pathToModelDescription::String: Contains the path to a file name that is selected to be read and converted to an XML document. In order to better extract the variables and meta data in the further process.

Returns

md::fmi2ModelDescription: Retuns a struct which provides the static information of ModelVariables.

Source

EzXML.jl

general information about the FMU

FMIImport.fmi2GetModelName — Function

fmi2GetModelName(md::fmi2ModelDescription)

Returns the tag 'modelName' from the model description.

Arguments

md::fmi2ModelDescription: Struct which provides the static information of ModelVariables.

Returns

md.modelName::String: Returns the tag 'modelName' from the model description.

function fmi2GetModelName(fmu::FMU2)

Returns the tag 'modelName' from the model description.

Arguments

fmu::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

Returns

fmu.modelDescription.modelName::String: Returns the tag 'modelName' from the model description.

FMIImport.fmi2GetGUID — Function

fmi2GetGUID(md::fmi2ModelDescription)

Returns the tag 'guid' from the model description.

Arguments

md::fmi2ModelDescription: Struct which provides the static information of ModelVariables.

Returns

md.guid::String: Returns the tag 'guid' from the model description.

fmi2GetGUID(fmu::FMU2)

Returns the tag 'guid' from the model description.

Arguments

fmu::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

Returns

fmu.modelDescription.guid::String: Returns the tag 'guid' from the model description.

FMIImport.fmi2IsCoSimulation — Function

fmi2IsCoSimulation(md::fmi2ModelDescription)

Returns true, if the FMU supports co simulation

Arguments

md::fmi2ModelDescription: Struct which provides the static information of ModelVariables.

Returns

::Bool: Returns true, if the FMU supports co simulation

fmi2IsCoSimulation(fmu::FMU2)

Returns true, if the FMU supports co simulation

Arguments

fmu::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

Returns

::Bool: Returns true, if the FMU supports co simulation

FMIImport.fmi2IsModelExchange — Function

fmi2IsModelExchange(md::fmi2ModelDescription)

Returns true, if the FMU supports model exchange

Arguments

md::fmi2ModelDescription: Struct which provides the static information of ModelVariables.

Returns

::Bool: Returns true, if the FMU supports model exchange

fmi2IsModelExchange(fmu::FMU2)

Returns true, if the FMU supports model exchange

Arguments

fmu::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

Returns

::Bool: Returns true, if the FMU supports model exchange

FMIImport.fmi2GetGenerationTool — Function

fmi2GetGenerationTool(md::fmi2ModelDescription)

Returns the tag 'generationtool' from the model description.

Arguments

md::fmi2ModelDescription: Struct which provides the static information of ModelVariables.

Returns

md.generationTool::Union{String, Nothing}: Returns the tag 'generationtool' from the model description.

fmi2GetGenerationTool(fmu::FMU2)

Returns the tag 'generationtool' from the model description.

Arguments

fmu::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

Returns

fmu.modelDescription.generationTool::Union{String, Nothing}: Returns the tag 'generationtool' from the model description.

FMIImport.fmi2GetGenerationDateAndTime — Function

fmi2GetGenerationDateAndTime(md::fmi2ModelDescription)

Returns the tag 'generationdateandtime' from the model description.

Arguments

md::fmi2ModelDescription: Struct which provides the static information of ModelVariables.

Returns

md.generationDateAndTime::DateTime: Returns the tag 'generationdateandtime' from the model description.

fmi2GetGenerationDateAndTime(fmu::FMU2)

Returns the tag 'generationdateandtime' from the model description.

Arguments

fmu::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

Returns

fmu.modelDescription.generationDateAndTime::DateTime: Returns the tag 'generationdateandtime' from the model description.

tecnical information about the FMU

FMIImport.fmi2GetModelIdentifier — Function

fmi2GetModelIdentifier(md::fmi2ModelDescription; type=nothing)

Returns the tag 'modelIdentifier' from CS or ME section.

Arguments

md::fmi2ModelDescription: Struct which provides the static information of ModelVariables.

Keywords

type=nothing: Defines whether a Co-Simulation or Model Exchange is present. (default = nothing)

Returns

md.modelExchange.modelIdentifier::String: Returns the tag modelIdentifier from ModelExchange section.

md.coSimulation.modelIdentifier::String: Returns the tag modelIdentifier from CoSimulation section.

FMIImport.fmi2GetVariableNamingConvention — Function

fmi2GetVariableNamingConvention(md::fmi2ModelDescription)

Returns the tag 'varaiblenamingconvention' from the model description.

Arguments

md::fmi2ModelDescription: Struct which provides the static information of ModelVariables.

Returns

md.variableNamingConvention::Union{fmi2VariableNamingConvention, Nothing}: Returns the tag 'variableNamingConvention' from the model description.

fmi2GetVariableNamingConvention(fmu::FMU2)

Returns the tag 'varaiblenamingconvention' from the model description.

Arguments

fmu::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

Returns

fmu.modelDescription.variableNamingConvention::Union{fmi2VariableNamingConvention, Nothing}: Returns the tag 'variableNamingConvention' from the model description.

FMICore.fmi2GetVersion — Function

fmi2GetVersion(fmu::FMU2)

Returns the version of the “fmi2Functions.h” header file which was used to compile the functions of the FMU.

Arguments

fmu::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

Returns

Returns a string from the address of a C-style (NUL-terminated) string. The string represents the version of the “fmi2Functions.h” header file which was used to compile the functions of the FMU. The function returns “fmiVersion” which is defined in this header file. The standard header file as documented in this specification has version “2.0”

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.22]: 2.1.4 Inquire Platform and Version Number of Header Files

FMISpec2.0.2[p.16]: 2.1.2 Platform Dependent Definitions

fmi2GetVersion(c::FMU2Component)

Returns the version of the “fmi2Functions.h” header file which was used to compile the functions of the FMU.

Arguments

c::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

Returns

Returns a string from the address of a C-style (NUL-terminated) string. The string represents the version of the “fmi2Functions.h” header file which was used to compile the functions of the FMU. The function returns “fmiVersion” which is defined in this header file. The standard header file as documented in this specification has version “2.0”

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.22]: 2.1.4 Inquire Platform and Version Number of Header Files

FMISpec2.0.2[p.16]: 2.1.2 Platform Dependent Definitions

Source: FMISpec2.0.2[p.22]: 2.1.4 Inquire Platform and Version Number of Header Files

Returns the version of the “fmi2Functions.h” header file which was used to compile the functions of the FMU. The function returns “fmiVersion” which is defined in this header file. The standard header file as documented in this specification has version “2.0”

FMICore.fmi2GetTypesPlatform — Function

fmi2GetTypesPlatform(fmu::FMU2)

Returns the string to uniquely identify the “fmi2TypesPlatform.h” header file used for compilation of the functions of the FMU. The standard header file, as documented in this specification, has fmi2TypesPlatform set to “default” (so this function usually returns “default”).

Arguments

fmu::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

Returns

Returns the string to uniquely identify the “fmi2TypesPlatform.h” header file used for compilation of the functions of the FMU.

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.22]: 2.1.4 Inquire Platform and Version Number of Header Files

FMISpec2.0.2[p.16]: 2.1.2 Platform Dependent Definitions

fmi2GetTypesPlatform(c::FMU2Component)

Returns the string to uniquely identify the “fmi2TypesPlatform.h” header file used for compilation of the functions of the FMU. The standard header file, as documented in this specification, has fmi2TypesPlatform set to “default” (so this function usually returns “default”).

Arguments

c::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

Returns

Returns the string to uniquely identify the “fmi2TypesPlatform.h” header file used for compilation of the functions of the FMU.

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.22]: 2.1.4 Inquire Platform and Version Number of Header Files

FMISpec2.0.2[p.16]: 2.1.2 Platform Dependent Definitions

Source: FMISpec2.0.2[p.22]: 2.1.4 Inquire Platform and Version Number of Header Files

Returns the string to uniquely identify the “fmi2TypesPlatform.h” header file used for compilation of the functions of the FMU. The standard header file, as documented in this specification, has fmi2TypesPlatform set to “default” (so this function usually returns “default”).

FMIImport.fmi2GetNumberOfEventIndicators — Function

fmi2GetNumberOfEventIndicators(md::fmi2ModelDescription)

Returns the tag 'numberOfEventIndicators' from the model description.

Arguments

md::fmi2ModelDescription: Struct which provides the static information of ModelVariables.

Returns

md.numberOfEventIndicators::Union{UInt, Nothing}: Returns the tag 'numberOfEventIndicators' from the model description.

fmi2GetNumberOfEventIndicators(fmu::FMU2)

Returns the tag 'numberOfEventIndicators' from the model description.

Arguments

fmu::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

Returns

fmu.modelDescription.numberOfEventIndicators::Union{UInt, Nothing}: Returns the tag 'numberOfEventIndicators' from the model description.

default experiment settings

FMIImport.fmi2GetDefaultStartTime — Function

fmi2GetDefaultStartTime(md::fmi2ModelDescription)

Returns startTime from DefaultExperiment if defined else defaults to nothing.

Arguments

md::fmi2ModelDescription: Struct which provides the static information of ModelVariables.

Returns

md.defaultExperiment.startTime::Union{Real,Nothing}: Returns a real value startTime from the DefaultExperiment if defined else defaults to nothing.

FMIImport.fmi2GetDefaultStopTime — Function

fmi2GetDefaultStopTime(md::fmi2ModelDescription)

Returns stopTime from DefaultExperiment if defined else defaults to nothing.

Arguments

md::fmi2ModelDescription: Struct which provides the static information of ModelVariables.

Returns

md.defaultExperiment.stopTime::Union{Real,Nothing}: Returns a real value stopTime from the DefaultExperiment if defined else defaults to nothing.

FMIImport.fmi2GetDefaultTolerance — Function

fmi2GetDefaultTolerance(md::fmi2ModelDescription)

Returns tolerance from DefaultExperiment if defined else defaults to nothing.

Arguments

md::fmi2ModelDescription: Struct which provides the static information of ModelVariables.

Returns

md.defaultExperiment.tolerance::Union{Real,Nothing}: Returns a real value tolerance from the DefaultExperiment if defined else defaults to nothing.

FMIImport.fmi2GetDefaultStepSize — Function

fmi2GetDefaultStepSize(md::fmi2ModelDescription)

Returns stepSize from DefaultExperiment if defined else defaults to nothing.

Arguments

md::fmi2ModelDescription: Struct which provides the static information of ModelVariables.

Returns

md.defaultExperiment.stepSize::Union{Real,Nothing}: Returns a real value setpSize from the DefaultExperiment if defined else defaults to nothing.

FMU capabilities

FMIImport.fmi2DependenciesSupported — Function

fmi2DependenciesSupported(md::fmi2ModelDescription)

Returns true if the FMU model description contains dependency information.

Arguments

md::fmi2ModelDescription: Struct which provides the static information of ModelVariables.

Returns

::Bool: Returns true, if the FMU model description contains dependency information.

FMIImport.fmi2DerivativeDependenciesSupported — Function

fmi2DerivativeDependenciesSupported(md::fmi2ModelDescription)

Returns if the FMU model description contains dependency information for derivatives.

Arguments

md::fmi2ModelDescription: Struct which provides the static information of ModelVariables.

Returns

::Bool: Returns true, if the FMU model description contains dependency information for derivatives.

FMIImport.fmi2CanGetSetState — Function

fmi2CanGetSetState(md::fmi2ModelDescription)

Returns true, if the FMU supports the getting/setting of states

Arguments

md::fmi2ModelDescription: Struct which provides the static information of ModelVariables.

Returns

::Bool: Returns true, if the FMU supports the getting/setting of states.

fmi2CanGetSetState(fmu::FMU2)

Returns true, if the FMU supports the getting/setting of states

Arguments

fmu::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

Returns

::Bool: Returns true, if the FMU supports the getting/setting of states.

FMIImport.fmi2CanSerializeFMUstate — Function

fmi2CanSerializeFMUstate(md::fmi2ModelDescription)

Returns true, if the FMU state can be serialized

Arguments

md::fmi2ModelDescription: Struct which provides the static information of ModelVariables.

Returns

::Bool: Returns true, if the FMU state can be serialized

fmi2CanSerializeFMUstate(fmu::FMU2)

Returns true, if the FMU state can be serialized

Arguments

fmu::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

Returns

::Bool: Returns true, if the FMU state can be serialized

FMIImport.fmi2ProvidesDirectionalDerivative — Function

fmi2ProvidesDirectionalDerivative(md::fmi2ModelDescription)

Returns true, if the FMU provides directional derivatives

Arguments

md::fmi2ModelDescription: Struct which provides the static information of ModelVariables.

Returns

::Bool: Returns true, if the FMU provides directional derivatives

fmi2ProvidesDirectionalDerivative(fmu::FMU2)

Returns true, if the FMU provides directional derivatives

Arguments

fmu::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

Returns

::Bool: Returns true, if the FMU provides directional derivatives

value references

FMIImport.fmi2GetValueReferencesAndNames — Function

fmi2GetValueReferencesAndNames(md::fmi2ModelDescription; vrs=md.valueReferences)

Returns a dictionary Dict(fmi2ValueReference, Array{String}) of value references and their corresponding names.

Arguments

md::fmi2ModelDescription: Struct which provides the static information of ModelVariables.

Keywords

vrs=md.valueReferences: Additional attribute valueReferences::Array{fmi2ValueReference} of the Model Description that is a handle to a (base type) variable value. Handle and base type uniquely identify the value of a variable. (default = md.valueReferences::Array{fmi2ValueReference})

Returns

dict::Dict{fmi2ValueReference, Array{String}}: Returns a dictionary that constructs a hash table with keys of type fmi2ValueReference and values of type Array{String}.

fmi2GetValueReferencesAndNames(fmu::FMU2)

Returns a dictionary Dict(fmi2ValueReference, Array{String}) of value references and their corresponding names.

Arguments

fmu::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

Returns

dict::Dict{fmi2ValueReference, Array{String}}: Returns a dictionary that constructs a hash table with keys of type fmi2ValueReference and values of type Array{String}.

FMIImport.fmi2GetNames — Function

fmi2GetNames(md::fmi2ModelDescription; vrs=md.valueReferences, mode=:first)

Returns a array of names corresponding to value references vrs.

Arguments

md::fmi2ModelDescription: Struct which provides the static information of ModelVariables.

Keywords

vrs=md.valueReferences: Additional attribute valueReferences::Array{fmi2ValueReference} of the Model Description that is a handle to a (base type) variable value. Handle and base type uniquely identify the value of a variable. (default = md.valueReferences::Array{fmi2ValueReference})

mode=:first: If there are multiple names per value reference, availabel modes are :first (default, pick only the first one), :group (pick all and group them into an array) and :flat (pick all, but flat them out into a 1D-array together with all other names)

Returns

names::Array{String}: Returns a array of names corresponding to value references vrs

fmi2GetNames(fmu::FMU2; vrs=md.valueReferences, mode=:first)

Returns a array of names corresponding to value references vrs.

Arguments

fmu::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

Keywords

vrs=md.valueReferences: Additional attribute valueReferences::Array{fmi2ValueReference} of the Model Description that is a handle to a (base type) variable value. Handle and base type uniquely identify the value of a variable. (default = md.valueReferences::Array{fmi2ValueReference})

mode=:first: If there are multiple names per value reference, availabel modes are :first (default, pick only the first one), :group (pick all and group them into an array) and :flat (pick all, but flat them out into a 1D-array together with all other names)

Returns

names::Array{String}: Returns a array of names corresponding to value references vrs

FMIImport.fmi2GetModelVariableIndices — Function

fmi2GetModelVariableIndices(md::fmi2ModelDescription; vrs=md.valueReferences)

Returns a array of indices corresponding to value references vrs

Arguments

md::fmi2ModelDescription: Struct which provides the static information of ModelVariables.

Keywords

vrs=md.valueReferences: Additional attribute valueReferences::Array{fmi2ValueReference} of the Model Description that is a handle to a (base type) variable value. Handle and base type uniquely identify the value of a variable. (default = md.valueReferences::Array{fmi2ValueReference})

Returns

names::Array{Integer}: Returns a array of indices corresponding to value references vrs

FMIImport.fmi2DataTypeForValueReference — Function

fmi2DataTypeForValueReference(md::fmi2ModelDescription, vr::fmi2ValueReference)

Returns the fmi2DataType (fmi2Real, fmi2Integer, fmi2Boolean, fmi2String) for a given Valuereference vr of a given FMU-ModelDescription md

In-/Outputs

FMIImport.fmi2GetInputValueReferencesAndNames — Function

fmi2GetInputValueReferencesAndNames(md::fmi2ModelDescription)

Returns a dict with (vrs, names of inputs).

Arguments

md::fmi2ModelDescription: Struct which provides the static information of ModelVariables.

fmu::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

Returns

dict::Dict{fmi2ValueReference, Array{String}}: Returns a dictionary that constructs a hash table with keys of type fmi2ValueReference and values of type Array{String}. So returns a dict with (vrs, names of inputs)

fmi2GetInputValueReferencesAndNames(fmu::FMU2)

Returns a dict with (vrs, names of inputs).

Arguments

fmu::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

Returns

dict::Dict{fmi2ValueReference, Array{String}}: Returns a dictionary that constructs a hash table with keys of type fmi2ValueReference and values of type Array{String}. So returns a dict with (vrs, names of inputs)

FMIImport.fmi2GetInputNames — Function

fmi2GetInputNames(md::fmi2ModelDescription; vrs=md.inputvalueReferences, mode=:first)

Returns names of inputs.

Arguments

md::fmi2ModelDescription: Struct which provides the static information of ModelVariables.

Keywords

vrs=md.inputvalueReferences: Additional attribute inputvalueReferences::Array{fmi2ValueReference} of the Model Description that is a handle to a (base type) variable value. Handle and base type uniquely identify the value of a variable. (default = md.valueReferences::Array{fmi2ValueReference})

mode=:first: If there are multiple names per value reference, availabel modes are :first (default, pick only the first one), :group (pick all and group them into an array) and :flat (pick all, but flat them out into a 1D-array together with all other names)

Returns

names::Array{String}: Returns a array of names corresponding to value references vrs

fmi2GetInputNames(fmu::FMU2; vrs=md.inputValueReferences, mode=:first)

Returns names of inputs.

Arguments

fmu::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

Keywords

vrs=md.inputvalueReferences: Additional attribute inputvalueReferences::Array{fmi2ValueReference} of the Model Description that is a handle to a (base type) variable value. Handle and base type uniquely identify the value of a variable. (default = md.valueReferences::Array{fmi2ValueReference})

mode=:first: If there are multiple names per value reference, availabel modes are :first (default, pick only the first one), :group (pick all and group them into an array) and :flat (pick all, but flat them out into a 1D-array together with all other names)

Returns

names::Array{String}: Returns a array of names corresponding to value references vrs

FMIImport.fmi2GetOutputValueReferencesAndNames — Function

fmi2GetOutputValueReferencesAndNames(md::fmi2ModelDescription)

Returns a dictionary Dict(fmi2ValueReference, Array{String}) of value references and their corresponding names.

Arguments

md::fmi2ModelDescription: Struct which provides the static information of ModelVariables.

Keywords

vrs=md.outputvalueReferences: Additional attribute outputvalueReferences::Array{fmi2ValueReference} of the Model Description that is a handle to a (base type) variable value. Handle and base type uniquely identify the value of a variable. (default = md.outputvalueReferences::Array{fmi2ValueReference})

Returns

dict::Dict{fmi2ValueReference, Array{String}}: Returns a dictionary that constructs a hash table with keys of type fmi2ValueReference and values of type Array{String}.So returns a dict with (vrs, names of outputs)

fmi2GetOutputValueReferencesAndNames(md::fmi2ModelDescription)

Returns a dictionary Dict(fmi2ValueReference, Array{String}) of value references and their corresponding names.

Arguments

fmu::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

Returns

dict::Dict{fmi2ValueReference, Array{String}}: Returns a dictionary that constructs a hash table with keys of type fmi2ValueReference and values of type Array{String}.So returns a dict with (vrs, names of outputs)

FMIImport.fmi2GetOutputNames — Function

fmi2GetOutputNames(md::fmi2ModelDescription; vrs=md.outputvalueReferences, mode=:first)

Returns names of outputs.

Arguments

md::fmi2ModelDescription: Struct which provides the static information of ModelVariables.

Keywords

vrs=md.outputvalueReferences: Additional attribute outputvalueReferences::Array{fmi2ValueReference} of the Model Description that is a handle to a (base type) variable value. Handle and base type uniquely identify the value of a variable. (default = md.outputvalueReferences::Array{fmi2ValueReference})

mode=:first: If there are multiple names per value reference, availabel modes are :first (default, pick only the first one), :group (pick all and group them into an array) and :flat (pick all, but flat them out into a 1D-array together with all other names)

Returns

names::Array{String}: Returns a array of names corresponding to value references vrs

fmi2GetOutputNames(fmu::FMU2; vrs=md.outputvalueReferences, mode=:first)

Returns names of outputs.

Arguments

fmu::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

Keywords

vrs=md.outputvalueReferences: Additional attribute outputvalueReferences::Array{fmi2ValueReference} of the Model Description that is a handle to a (base type) variable value. Handle and base type uniquely identify the value of a variable. (default = md.outputvalueReferences::Array{fmi2ValueReference})

mode=:first: If there are multiple names per value reference, availabel modes are :first (default, pick only the first one), :group (pick all and group them into an array) and :flat (pick all, but flat them out into a 1D-array together with all other names)

Returns

names::Array{String}: Returns a array of names corresponding to value references vrs

Parameters

FMIImport.fmi2GetParameterValueReferencesAndNames — Function

fmi2GetParameterValueReferencesAndNames(md::fmi2ModelDescription)

Returns a dictionary Dict(fmi2ValueReference, Array{String}) of parameterValueReferences and their corresponding names.

Arguments

md::fmi2ModelDescription: Struct which provides the static information of ModelVariables.

Returns

dict::Dict{fmi2ValueReference, Array{String}}: Returns a dictionary that constructs a hash table with keys of type fmi2ValueReference and values of type Array{String}. So returns a dict with (vrs, names of parameters).

See also fmi2GetValueReferencesAndNames.

fmi2GetParameterValueReferencesAndNames(fmu::FMU2)

Returns a dictionary Dict(fmi2ValueReference, Array{String}) of parameterValueReferences and their corresponding names.

Arguments

fmu::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

Returns

dict::Dict{fmi2ValueReference, Array{String}}: Returns a dictionary that constructs a hash table with keys of type fmi2ValueReference and values of type Array{String}. So returns a dict with (vrs, names of parameters).

See also fmi2GetValueReferencesAndNames.

FMIImport.fmi2GetParameterNames — Function

fmi2GetParameterNames(md::fmi2ModelDescription; vrs=md.parameterValueReferences, mode=:first)

Returns names of parameters.

Arguments

md::fmi2ModelDescription: Struct which provides the static information of ModelVariables.

Keywords

vrs=md.parameterValueReferences: Additional attribute parameterValueReferences::Array{fmi2ValueReference} of the Model Description that is a handle to a (base type) variable value. Handle and base type uniquely identify the value of a variable. (default = md.parameterValueReferences::Array{fmi2ValueReference})

mode=:first: If there are multiple names per value reference, availabel modes are :first (default, pick only the first one), :group (pick all and group them into an array) and :flat (pick all, but flat them out into a 1D-array together with all other names)

Returns

names::Array{String}: Returns a array of names corresponding to parameter value references vrs

fmi2GetParameterNames(fmu::FMU2; vrs=md.parameterValueReferences, mode=:first)

Returns names of parameters.

Arguments

fmu::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

Keywords

vrs=md.parameterValueReferences: Additional attribute parameterValueReferences::Array{fmi2ValueReference} of the Model Description that is a handle to a (base type) variable value. Handle and base type uniquely identify the value of a variable. (default = md.parameterValueReferences::Array{fmi2ValueReference})

mode=:first: If there are multiple names per value reference, availabel modes are :first (default, pick only the first one), :group (pick all and group them into an array) and :flat (pick all, but flat them out into a 1D-array together with all other names)

Returns

names::Array{String}: Returns a array of names corresponding to parameter value references vrs

States

FMIImport.fmi2GetNumberOfStates — Function

fmi2GetNumberOfStates(md::fmi2ModelDescription)

Returns the number of states of the FMU.

Arguments

md::fmi2ModelDescription: Struct which provides the static information of ModelVariables.

Returns

Returns the length of the md.valueReferences::Array{fmi2ValueReference} corresponding to the number of states of the FMU.

fmi2GetNumberOfStates(fmu::FMU2)

Returns the number of states of the FMU.

Arguments

fmu::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

Returns

Returns the length of the fmu.modelDescription.valueReferences::Array{fmi2ValueReference} corresponding to the number of states of the FMU.

FMIImport.fmi2GetStateValueReferencesAndNames — Function

fmi2GetStateValueReferencesAndNames(md::fmi2ModelDescription)

Returns a dictionary Dict(fmi2ValueReference, Array{String}) of state value references and their corresponding names.

Arguments

md::fmi2ModelDescription: Struct which provides the static information of ModelVariables.

Returns

dict::Dict{fmi2ValueReference, Array{String}}: Returns a dictionary that constructs a hash table with keys of type fmi2ValueReference and values of type Array{String}. So returns a dict with (vrs, names of states)

fmi2GetStateValueReferencesAndNames(fmu::FMU2)

Returns dict(vrs, names of states).

Returns a dictionary Dict(fmi2ValueReference, Array{String}) of state value references and their corresponding names.

Arguments

fmu::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

Returns

dict::Dict{fmi2ValueReference, Array{String}}: Returns a dictionary that constructs a hash table with keys of type fmi2ValueReference and values of type Array{String}. So returns a dict with (vrs, names of states)

FMIImport.fmi2GetStateNames — Function

fmi2GetStateNames(fmu::FMU2; vrs=md.stateValueReferences, mode=:first)

Returns names of states.

Arguments

md::fmi2ModelDescription: Struct which provides the static information of ModelVariables.

Keywords

vrs=md.stateValueReferences: Additional attribute parameterValueReferences::Array{fmi2ValueReference} of the Model Description that is a handle to a (base type) variable value. Handle and base type uniquely identify the value of a variable. (default = md.stateValueReferences::Array{fmi2ValueReference})

mode=:first: If there are multiple names per value reference, availabel modes are :first (default, pick only the first one), :group (pick all and group them into an array) and :flat (pick all, but flat them out into a 1D-array together with all other names)

Returns

names::Array{String}: Returns a array of names corresponding to parameter value references vrs

fmi2GetStateNames(fmu::FMU2; vrs=md.stateValueReferences, mode=:first)

Returns names of states.

Arguments

fmu::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

Keywords

vrs=md.stateValueReferences: Additional attribute parameterValueReferences::Array{fmi2ValueReference} of the Model Description that is a handle to a (base type) variable value. Handle and base type uniquely identify the value of a variable. (default = md.stateValueReferences::Array{fmi2ValueReference})

mode=:first: If there are multiple names per value reference, availabel modes are :first (default, pick only the first one), :group (pick all and group them into an array) and :flat (pick all, but flat them out into a 1D-array together with all other names)

Returns

names::Array{String}: Returns a array of names corresponding to parameter value references vrs

Derivatives

FMIImport.fmi2GetDerivateValueReferencesAndNames — Function

fmi2GetDerivateValueReferencesAndNames(md::fmi2ModelDescription)

Returns a dictionary Dict(fmi2ValueReference, Array{String}) of derivative value references and their corresponding names.

Arguments

md::fmi2ModelDescription: Struct which provides the static information of ModelVariables.

Returns

dict::Dict{fmi2ValueReference, Array{String}}: Returns a dictionary that constructs a hash table with keys of type fmi2ValueReference and values of type Array{String}. So returns a dict with (vrs, names of derivatives)

See also fmi2GetValueReferencesAndNames

fmi2GetDerivateValueReferencesAndNames(fmu::FMU2)

Returns a dictionary Dict(fmi2ValueReference, Array{String}) of derivative value references and their corresponding names.

Arguments

fmu::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

Returns

dict::Dict{fmi2ValueReference, Array{String}}: Returns a dictionary that constructs a hash table with keys of type fmi2ValueReference and values of type Array{String}. So returns a dict with (vrs, names of derivatives)

See also fmi2GetValueReferencesAndNames

FMIImport.fmi2GetDerivativeNames — Function

fmi2GetDerivativeNames(md::fmi2ModelDescription; vrs=md.derivativeValueReferences, mode=:first)

Returns names of derivatives.

Arguments

md::fmi2ModelDescription: Struct which provides the static information of ModelVariables.

Keywords

vrs=md.derivativeValueReferences: Additional attribute derivativeValueReferences::Array{fmi2ValueReference} of the Model Description that is a handle to a (base type) variable value. Handle and base type uniquely identify the value of a variable. (default = md.derivativeValueReferences::Array{fmi2ValueReference})

mode=:first: If there are multiple names per value reference, availabel modes are :first (default, pick only the first one), :group (pick all and group them into an array) and :flat (pick all, but flat them out into a 1D-array together with all other names)

Returns

names::Array{String}: Returns a array of names corresponding to parameter value references vrs

fmi2GetDerivativeNames(fmu::FMU2; vrs=md.derivativeValueReferences, mode=:first)

Returns names of derivatives.

Arguments

fmu::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

Keywords

vrs=md.derivativeValueReferences: Additional attribute derivativeValueReferences::Array{fmi2ValueReference} of the Model Description that is a handle to a (base type) variable value. Handle and base type uniquely identify the value of a variable. (default = md.derivativeValueReferences::Array{fmi2ValueReference})

mode=:first: If there are multiple names per value reference, availabel modes are :first (default, pick only the first one), :group (pick all and group them into an array) and :flat (pick all, but flat them out into a 1D-array together with all other names)

Returns

names::Array{String}: Returns a array of names corresponding to parameter value references vrs

Variables

FMIImport.fmi2GetNamesAndDescriptions — Function

fmi2GetNamesAndDescriptions(md::fmi2ModelDescription)

Returns a dictionary of variables with their descriptions.

Arguments

md::fmi2ModelDescription: Struct which provides the static information of ModelVariables.

Returns

dict::Dict{String, String}: Returns a dictionary that constructs a hash table with keys of type String and values of type String. So returns a dict with ( md.modelVariables[i].name::String, md.modelVariables[i].description::Union{String, Nothing}). (Creates a tuple (name, description) for each i in 1:length(md.modelVariables))

fmi2GetNamesAndDescriptions(fmu::FMU2)

Returns a dictionary of variables with their descriptions.

Arguments

fmu::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

Returns

dict::Dict{String, String}: Returns a dictionary that constructs a hash table with keys of type String and values of type String. So returns a dict with ( md.modelVariables[i].name::String, md.modelVariables[i].description::Union{String, Nothing}). (Creates a tuple (name, description) for each i in 1:length(md.modelVariables))

FMIImport.fmi2GetNamesAndUnits — Function

fmi2GetNamesAndUnits(md::fmi2ModelDescription)

Returns a dictionary of variables with their units.

Arguments

md::fmi2ModelDescription: Struct which provides the static information of ModelVariables.

Returns

dict::Dict{String, String}: Returns a dictionary that constructs a hash table with keys of type String and values of type String. So returns a dict with ( md.modelVariables[i].name::String, md.modelVariables[i]._Real.unit::Union{String, Nothing}). (Creates a tuple (name, unit) for each i in 1:length(md.modelVariables))

See also fmi2GetUnit.

fmi2GetNamesAndUnits(fmu::FMU2)

Returns a dictionary of variables with their units.

Arguments

fmu::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

Returns

dict::Dict{String, String}: Returns a dictionary that constructs a hash table with keys of type String and values of type String. So returns a dict with ( md.modelVariables[i].name::String, md.modelVariables[i]._Real.unit::Union{String, Nothing}). (Creates a tuple (name, unit) for each i in 1:length(md.modelVariables))

See also fmi2GetUnit.

FMIImport.fmi2GetNamesAndInitials — Function

fmi2GetNamesAndInitials(md::fmi2ModelDescription)

Returns a dictionary of variables with their initials.

Arguments

md::fmi2ModelDescription: Struct which provides the static information of ModelVariables.

Returns

dict::Dict{String, Cuint}: Returns a dictionary that constructs a hash table with keys of type String and values of type Cuint. So returns a dict with ( md.modelVariables[i].name::String, md.modelVariables[i].inital::Union{fmi2Initial, Nothing}). (Creates a tuple (name,initial) for each i in 1:length(md.modelVariables))

See also fmi2GetInitial.

fmi2GetNamesAndInitials(fmu::FMU2)

Returns a dictionary of variables with their initials.

Arguments

fmu::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

Returns

dict::Dict{String, Cuint}: Returns a dictionary that constructs a hash table with keys of type String and values of type Cuint. So returns a dict with ( md.modelVariables[i].name::String, md.modelVariables[i].inital::Union{fmi2Initial, Nothing}). (Creates a tuple (name,initial) for each i in 1:length(md.modelVariables))

See also fmi2GetInitial.

FMIImport.fmi2GetInputNamesAndStarts — Function

fmi2GetInputNamesAndStarts(md::fmi2ModelDescription)

Returns a dictionary of input variables with their starting values.

Arguments

md::fmi2ModelDescription: Struct which provides the static information of ModelVariables.

Returns

dict::Dict{String, Array{fmi2ValueReferenceFormat}}: Returns a dictionary that constructs a hash table with keys of type String and values of type fmi2ValueReferenceFormat. So returns a dict with ( md.modelVariables[i].name::String, starts:: Array{fmi2ValueReferenceFormat} ). (Creates a tuple (name, starts) for each i in inputIndices)

See also fmi2GetStartValue.

fmi2GetInputNamesAndStarts(fmu::FMU2)

Returns a dictionary of input variables with their starting values.

Arguments

fmu::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

Returns

dict::Dict{String, Array{fmi2ValueReferenceFormat}}: Returns a dictionary that constructs a hash table with keys of type String and values of type fmi2ValueReferenceFormat. So returns a dict with ( md.modelVariables[i].name::String, starts:: Array{fmi2ValueReferenceFormat} ). (Creates a tuple (name, starts) for each i in inputIndices)

See also fmi2GetStartValue.