FMI for Co-Simulation

This chapter defines the Functional Mock-up Interface (FMI) for the coupling of two or more simulation models in a Co-Simulation environment (FMI for Co-Simulation). Co-Simulation is a rather general approach to the simulation of coupled technical systems and coupled physical phenomena in engineering with focus on instationary (time-dependent) problems.

Transfer of Input / Output Values and Parameters

In order to enable the slave to interpolate the continuous real inputs between communication steps, the derivatives of the inputs with respect to time can be provided. Also, higher derivatives can be set to allow higher order interpolation.

FMIImport.fmi3CallbackIntermediateUpdate — Function

fmi3CallbackIntermediateUpdate(instanceEnvironment::Ptr{Cvoid},
    intermediateUpdateTime::fmi3Float64,
    intermediateVariableSetRequested::fmi3Boolean,
    intermediateVariableGetAllowed::fmi3Boolean,
    intermediateStepFinished::fmi3Boolean,
    canReturnEarly::fmi3Boolean,
    earlyReturnRequested::Ptr{fmi3Boolean},
    earlyReturnTime::Ptr{fmi3Float64})

When a Co-Simulation FMU provides values for its output variables at intermediate points between two consecutive communication points, and is able to receive new values for input variables at these intermediate points, the Intermediate Update Callback function is called. This is typically required when the FMU uses a numerical solver to integrate the FMU's internal state between communication points in fmi3DoStep. The callback function switches the FMU from Step Mode (see 4.2.1.) in the Intermediate Update Mode (see 4.2.2.) and returns to Step Mode afterwards. If the ModelDescription has the "providesIntermediateUpdate" flag, the Intermediate update callback function is called. That flag is ignored in ModelExchange and ScheduledExecution.

Arguments

instanceEnvironment::Ptr{FMU3InstanceEnvironment}: is the instance name of the model that calls this function.

intermediateUpdateTime::fmi3Float64: is the internal value of the independent variable [typically simulation time] of the FMU at which the callback has been called for intermediate and final steps. If an event happens or an output Clock ticks, intermediateUpdateTime is the time of event or output Clock tick. In Co-Simulation, intermediateUpdateTime is restricted by the arguments to fmi3DoStep as follows: currentCommunicationPoint ≤ intermediateUpdateTime ≤ (currentCommunicationPoint + communicationStepSize). The FMU must not call the callback function fmi3CallbackIntermediateUpdate with an intermediateUpdateTime that is smaller than the intermediateUpdateTime given in a previous call of fmi3CallbackIntermediateUpdate with intermediateStepFinished == fmi3True.

intermediateVariableSetRequested::fmi3Boolean: If intermediateVariableSetRequested == fmi3True, the co-simulation algorithm may provide intermediate values for continuous input variables with intermediateUpdate = true by calling fmi3Set{VariableType}. The set of variables for which the co-simulation algorithm will provide intermediate values is declared through the requiredIntermediateVariables argument to fmi3InstantiateXXX. If a co-simulation algorithm does not provide a new value for any of the variables contained in the set it registered, the last value set remains.

intermediateVariableGetAllowed::fmi3Boolean: If intermediateVariableGetAllowed == fmi3True, the co-simulation algorithm may collect intermediate output variables by calling fmi3Get{VariableType} for variables with intermediateUpdate = true. The set of variables for which the co-simulation algorithm can get values is supplied through the requiredIntermediateVariables argument to fmi3InstantiateXXX.

intermediateStepFinished::fmi3Boolean: If intermediateStepFinished == fmi3False, the intermediate outputs of the FMU that the co-simulation algorithm inquires with fmi3Get{VariableType} resulting from tentative internal solver states and may still change for the same intermediateUpdateTime [e.g., if the solver deems the tentative state to cause a too high approximation error, it may go back in time and try to re-estimate the state using smaller internal time steps]. If intermediateStepFinished == fmi3True, intermediate outputs inquired by the co-simulation algorithm with fmi3Get{VariableType} correspond to accepted internal solver step.

canReturnEarly::fmi3Boolean: When canReturnEarly == fmi3True the FMU signals to the co-simulation algorithm its ability to return early from the current fmi3DoStep.

earlyReturnRequested::Ptr{fmi3Boolean}: If and only if canReturnEarly == fmi3True, the co-simulation algorithm may request the FMU to return early from fmi3DoStep by setting earlyReturnRequested == fmi3True.

earlyReturnTime::Ptr{fmi3Float64}: is used to signal the FMU at which time to return early from the current fmi3DoStep, if the return value of earlyReturnRequested == fmi3True. If the earlyReturnTime is greater than the last signaled intermediateUpdateTime, the FMU may integrate up to the time instant earlyReturnTime.

Returns

nothing

Source

FMISpec3.0 Link: https://fmi-standard.org/

FMISpec3.0: 4.2.2. State: Intermediate Update Mode

Computation

The computation of time steps is controlled by the following function.

FMICore.fmi3EnterStepMode — Function

fmi3EnterStepMode(c::FMU3Instance; soft::Bool=false)

This function must be called to change from Event Mode into Step Mode in Co-Simulation (see 4.2.).

Arguments

c::FMU3Instance: Mutable struct represents an instantiated instance of an FMU in the FMI 3.0 Standard.

Keywords

soft::Bool=false: If the Keyword soft = true the fmi3Teminate needs to be called in state fmi3InstanceStateContinuousTimeMode or fmi3InstanceStateEventMode.

Returns

status::fmi3Status: Return status is an enumeration of type fmi3Status and indicates the success of the function call.

More detailed:

fmi3OK: all well

fmi3Warning: things are not quite right, but the computation can continue

fmi3Discard: if the slave computed successfully only a subinterval of the communication step

fmi3Error: the communication step could not be carried out at all

fmi3Fatal: if an error occurred which corrupted the FMU irreparably

Source

FMISpec3.0 Link: https://fmi-standard.org/

FMISpec3.0: 2.2.3 Platform Dependent Definitions

FMISpec3.0: 2.2.4 Status Returned by Functions

FMISpec3.0: 2.3.5. State: Event Mode

See also fmi3EnterStepMode.

Source: FMISpec3.0, Version D5ef1c1: 2.3.5. State: Event Mode

This function must be called to change from Event Mode into Step Mode in Co-Simulation (see 4.2.).

fmi3EnterStepMode(fmu::FMU3)

Wrapper for fmi3EnterStepMode(c::FMU3Instance) without a provided FMU3Instance. (Instance c gets selected from fmu)

FMICore.fmi3DoStep! — Function

fmi3DoStep!(c::FMU3Instance, currentCommunicationPoint::fmi3Float64, communicationStepSize::fmi3Float64, noSetFMUStatePriorToCurrentPoint::fmi3Boolean,
                eventEncountered::Ref{fmi3Boolean}, terminateSimulation::Ref{fmi3Boolean}, earlyReturn::Ref{fmi3Boolean}, lastSuccessfulTime::Ref{fmi3Float64})

The computation of a time step is started.

TODO argmuents

Arguments

c::FMU3Instance: Mutable struct represents an instantiated instance of an FMU in the FMI 3.0 Standard.

currentCommunicationPoint::fmi3Float64:

communicationStepSize::fmi3Float64:

noSetFMUStatePriorToCurrentPoint::fmi3Boolean:

eventEncountered::Ref{fmi3Boolean}:

terminateSimulation::Ref{fmi3Boolean}:

earlyReturn::Ref{fmi3Boolean}:

lastSuccessfulTime::Ref{fmi3Float64}:

Returns

status::fmi3Status: Return status is an enumeration of type fmi3Status and indicates the success of the function call.

More detailed:

fmi3OK: all well

fmi3Warning: things are not quite right, but the computation can continue

fmi3Discard: if the slave computed successfully only a subinterval of the communication step

fmi3Error: the communication step could not be carried out at all

fmi3Fatal: if an error occurred which corrupted the FMU irreparably

Source

FMISpec3.0 Link: https://fmi-standard.org/

FMISpec3.0: 2.2.3 Platform Dependent Definitions

FMISpec3.0: 2.2.4 Status Returned by Functions

FMISpec3.0: 4.2.1. State: Step Mode

See also fmi3DoStep!.

fmi3DoStep!(c::FMU3Instance, currentCommunicationPoint::Union{Real, Nothing} = nothing, communicationStepSize::Union{Real, Nothing} = nothing, noSetFMUStatePriorToCurrentPoint::Bool = true,
    eventEncountered::fmi3Boolean = fmi3False, terminateSimulation::fmi3Boolean = fmi3False, earlyReturn::fmi3Boolean = fmi3False, lastSuccessfulTime::fmi3Float64 = 0.0)

The computation of a time step is started.

TODO argmuents

Arguments

c::FMU3Instance: Mutable struct represents an instantiated instance of an FMU in the FMI 3.0 Standard.

currentCommunicationPoint::Union{Real, Nothing} = nothing

communicationStepSize::Union{Real, Nothing} = nothing

noSetFMUStatePriorToCurrentPoint::Bool = true

eventEncountered::fmi3Boolean = fmi3False

terminateSimulation::fmi3Boolean = fmi3False

earlyReturn::fmi3Boolean = fmi3False

lastSuccessfulTime::fmi3Float64 = 0.0

Returns

status::fmi3Status: Return status is an enumeration of type fmi3Status and indicates the success of the function call.

More detailed:

fmi3OK: all well

fmi3Warning: things are not quite right, but the computation can continue

fmi3Discard: if the slave computed successfully only a subinterval of the communication step

fmi3Error: the communication step could not be carried out at all

fmi3Fatal: if an error occurred which corrupted the FMU irreparably

Source

FMISpec3.0 Link: https://fmi-standard.org/

FMISpec3.0: 2.2.3 Platform Dependent Definitions

FMISpec3.0: 2.2.4 Status Returned by Functions

FMISpec3.0: 4.2.1. State: Step Mode

See also fmi3DoStep!.

Source: FMISpec3.0, Version D5ef1c1: 4.2.1. State: Step Mode

The computation of a time step is started.

fmi3DoStep!(fmu::FMU3, args...; kwargs...)

Wrapper for fmi3DoStep!(c::FMU3Instance, args...; kwargs...) without a provided FMU3Instance. (Instance c gets selected from fmu)

Retrieving Status Information from the Slave

Status information is retrieved from the slave by the following functions: