Simuation results

FMI.fmi2SimulationResultGetValuesAtIndex — Function

collects all data points for variable save index ´i´ inside a fmi2SimulationResult ´sd´

FMI.fmi2SimulationResultGetTime — Function

collects all time data points inside a fmi2SimulationResult ´sd´

FMI.fmi2SimulationResultGetValues — Function

collects all data points for variable with value reference ´tvr´ inside a fmi2SimulationResult ´sd´

collects all data points for variable with value name ´s´ inside a fmi2SimulationResult ´sd´

Parsing variable names to ValueReferences

FMI.fmi2String2ValueReference — Function

Returns the ValueReference coresponding to the variable name

Returns an array of ValueReferences coresponding to the variable names

FMI.fmi2ValueReference2String — Function

Returns an array of variable names matching a fmi2ValueReference

Opening and closing FMUs

FMI.fmi2Unzip — Function

create a copy of the .fmu file as a .zip folder and unzips it returns the paths to the ziped and ziped folders

FMI.fmi2Load — Function

sets the properties of the fmu by reading the modelDescription.xml retrieve all the pointers of binary functions for later @userplot

returns the instance of the fmu struct

FMI.fmi2Unload — Function

Unload a fmu under the FMI-Standard 2.0.2

Free the allocated memory, close the binaries and remove temporary zip and unziped fmu fmi2ModelDescription

Simulate FMU

FMI.fmi2Simulate — Function

Starts a simulation of the fmu instance for the matching fmu type, if both types are available the CoSimulation Simulation is started

Starts a simulation of the fmu instance for the matching fmu type, if both types are available the CoSimulation Simulation is started

FMI.fmi2SimulateCS — Function

Starts a simulation of the CoSimulation FMU instance

Starts a simulation of the CoSimulation FMU instance

Starts a simulation of the CoSimulation FMU instance

Starts a simulation of the CoSimulation FMU instance

FMI.fmi2SimulateME — Function

Source: FMISpec2.0.2[p.90 ff]: 3.2.4 Pseudocode Example simulates an fmu instance for the set simulation time