Library Functions

Many of the functions in this library are based on already defined functions of the FMIImport.jl library. Within this library, a distinction is made between the FMI-standards (FMI2 and FMI3), which is why the links to both function collections are provided in the following.

FMI2Standard: https://thummeto.github.io/FMIImport.jl/dev/fmi2_library/

FMI3Standard: https://thummeto.github.io/FMIImport.jl/dev/fmi3_library/

In this collection only the essential and not in the standard defined functions are listed.

Simulate FMUs

FMI.fmiSimulate — Function

fmiSimulate(str::Union{fmi2Struct, fmi3Struct}, t_start::Union{Real, Nothing} = nothing, t_stop::Union{Real, Nothing} = nothing;
            tolerance::Union{Real, Nothing} = nothing,
            dt::Union{Real, Nothing} = nothing,
            solver = nothing,
            customFx = nothing,
            recordValues::fmi2ValueReferenceFormat = nothing,
            saveat = [],
            setup::Bool = true,
            reset::Union{Bool, Nothing} = nothing, # nothing = auto
            inputValueReferences::fmi2ValueReferenceFormat = nothing,
            inputFunction = nothing,
            parameters::Union{Dict{<:Any, <:Any}, Nothing} = nothing,
            dtmax::Union{Real, Nothing} = nothing,
            kwargs...)

Starts a simulation of the FMU instance for the matching FMU type, if both types are available, CS is preferred.

Arguments

str::Union{fmi2StructMD, fmi3StructMD}: Representative for an FMU in the FMI 2.0.2 Standard or FMI 3.0 Standard. Other notation:

More detailed: fmi2Struct = Union{FMU2, FMU2Component} More detailed: fmi3Struct = Union{FMU3, FMU3Instance}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

str::FMU3: Mutable struct representing an FMU in the FMI 3.0 Standard.

str::FMU3Instance: Mutable struct represents a pointer to an FMU specific data structure that contains the information needed. Also in FMI 3.0 Standard.

t_start::Union{Real, Nothing} = nothing: Set the start time to a value of type Real or the default value from the model description is used.

t_stop::Union{Real, Nothing} = nothing: Set the end time to a value of type Real or the default value from the model description is used.

Keywords

tolerance::Union{Real, Nothing} = nothing: Real number to set the tolerance for any OED-solver

dt::Union{Real, Nothing} = nothing: Real number to set the step size of the OED-solver. Defaults to an automatic choice if the method is adaptive. More Info: DifferentialEquations.jl Documentation

solver = nothing: Any Julia-supported OED-solver (default is Tsit5). More Info: DifferentialEquations.jl Documentation

customFx = nothing: [deprecated] Ability to give a custom state derivative function ẋ=f(x,t)

recordValues::fmi2ValueReferenceFormat = nothing: AbstractArray of variables (strings or variableIdentifiers) to record. Results are returned as DiffEqCallbacks.SavedValues

saveat = []: Time points to save values at (interpolated). More Info: DifferentialEquations.jl Documentation

setup::Bool = true: Boolean, if FMU should be setup (default: setup=true)

reset::Union{Bool, Nothing} = nothing: Boolean, if FMU should be reset before simulation (default: reset:=auto)

inputValueReferences::fmi2ValueReferenceFormat = nothing: AbstractArray of input variables (strings or variableIdentifiers) to set at every simulation step

inputFunction = nothing: Function to retrieve the values to set the inputs to

parameters::Union{Dict{<:Any, <:Any}, Nothing} = nothing: Dictionary of parameter variables (strings or variableIdentifiers) and values (Real, Integer, Boolean, String) to set parameters during initialization

dtmax::Union{Real, Nothing} = nothing: Real number for setting maximum dt for adaptive timestepping for the ODE solver. The default values are package dependent. More Info: DifferentialEquations.jl Documentation

kwargs...: Further parameters of already defined functions solve(args..., kwargs...) from the library DifferentialEquations.jl

Returns

success::Bool for CS-FMUs

ODESolution for ME-FMUs

if keyword recordValues is set, a tuple of type (success::Bool, DiffEqCallbacks.SavedValues) for CS-FMUs

if keyword recordValues is set, a tuple of type (ODESolution, DiffEqCallbacks.SavedValues) for ME-FMUs

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.22]: 2.1.4 Inquire Platform and Version Number of Header Files

FMISpec2.0.2[p.16]: 2.1.2 Platform Dependent Definitions

See also fmi2Simulate, fmi2SimulateME, fmi2SimulateCS, fmi2Struct, FMU2, FMU2Component.

FMI.fmiSimulateCS — Function

fmiSimulateCS(str::Union{fmi2Struct,fmi3Struct}, t_start::Union{Real, Nothing} = nothing, t_stop::Union{Real, Nothing} = nothing;
            tolerance::Union{Real, Nothing} = nothing,
            dt::Union{Real, Nothing} = nothing,
            solver = nothing,
            customFx = nothing,
            recordValues::fmi2ValueReferenceFormat = nothing,
            saveat = [],
            setup::Bool = true,
            reset::Union{Bool, Nothing} = nothing, # nothing = auto
            inputValueReferences::fmi2ValueReferenceFormat = nothing,
            inputFunction = nothing,
            parameters::Union{Dict{<:Any, <:Any}, Nothing} = nothing,
            dtmax::Union{Real, Nothing} = nothing,
            kwargs...)

Starts a simulation of the Co-Simulation FMU instance.

Arguments

str::Union{fmi2StructMD, fmi3StructMD}: Representative for an FMU in the FMI 2.0.2 Standard or FMI 3.0 Standard. Other notation:

More detailed: fmi2Struct = Union{FMU2, FMU2Component} More detailed: fmi3Struct = Union{FMU3, FMU3Instance}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

str::FMU3: Mutable struct representing an FMU in the FMI 3.0 Standard.

str::FMU3Instance: Mutable struct represents a pointer to an FMU specific data structure that contains the information needed. Also in FMI 3.0 Standard.

t_start::Union{Real, Nothing} = nothing: Set the start time to a value of type Real or the default value from the model description is used.

t_stop::Union{Real, Nothing} = nothing: Set the end time to a value of type Real or the default value from the model description is used.

Keywords

tolerance::Union{Real, Nothing} = nothing: Real number to set the tolerance for any OED-solver

dt::Union{Real, Nothing} = nothing: Real number to set the step size of the OED-solver. Defaults to an automatic choice if the method is adaptive. More Info: DifferentialEquations.jl Documentation

solver = nothing: Any Julia-supported OED-solver (default is Tsit5). More Info: DifferentialEquations.jl Documentation

customFx = nothing: [deprecated] Ability to give a custom state derivative function ẋ=f(x,t)

recordValues::fmi2ValueReferenceFormat = nothing: AbstractArray of variables (strings or variableIdentifiers) to record. Results are returned as DiffEqCallbacks.SavedValues

saveat = []: Time points to save values at (interpolated). More Info: DifferentialEquations.jl Documentation

setup::Bool = true: Boolean, if FMU should be setup (default: setup=true)

reset::Union{Bool, Nothing} = nothing: Boolean, if FMU should be reset before simulation (default: reset:=auto)

inputValueReferences::fmi2ValueReferenceFormat = nothing: AbstractArray of input variables (strings or variableIdentifiers) to set at every simulation step

inputFunction = nothing: Function to retrieve the values to set the inputs to

parameters::Union{Dict{<:Any, <:Any}, Nothing} = nothing: Dictionary of parameter variables (strings or variableIdentifiers) and values (Real, Integer, Boolean, String) to set parameters during initialization

dtmax::Union{Real, Nothing} = nothing: Real number for setting maximum dt for adaptive timestepping for the ODE solver. The default values are package dependent. More Info: DifferentialEquations.jl Documentation

kwargs...: Further parameters of already defined functions solve(args..., kwargs...) from the library DifferentialEquations.jl

Returns

If keyword recordValues is not set, a boolean success is returned (simulation success).

If keyword recordValues is set, a tuple of type (true, DiffEqCallbacks.SavedValues) or (false, nothing).

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.22]: 2.1.4 Inquire Platform and Version Number of Header Files

FMISpec2.0.2[p.16]: 2.1.2 Platform Dependent Definitions

See also fmi2SimulateCS, fmi2Simulate, fmi2SimulateME, fmi2Struct, FMU2, FMU2Component, fmi3SimulateCS, fmi3Simulate, fmi3SimulateME, fmi3Struct, FMU3, FMU3Instance.

Missing docstring.

Missing docstring for fmiSimulateME. Check Documenter's build log for details.

FMI.fmiLoad — Function

fmiLoad(pathToFMU::String; unpackPath=nothing, type=nothing)

Load FMUs independent of the FMI version, currently supporting version 2.0.X and 3.0.

Arguments

pathToFMU::String: String that contains the paths of ziped and unziped FMU folders.

Keywords

unpackPath=nothing: Via optional argument unpackPath, a path to unpack the FMU can be specified (default: system temporary directory).

type::Union{CS, ME, SE} = nothing: Via type, a FMU type can be selected. If none of the unified type set is used, the default value type = nothing will be used.

Returns

Returns the instance of the FMU struct.

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.22]: 2.1.4 Inquire Platform and Version Number of Header Files

FMISpec2.0.2[p.16]: 2.1.2 Platform Dependent Definitions

See also fmi2Load, fmi3Load.

FMI.fmiUnload — Function

fmiUnload(fmu::Union{FMU2, FMU3})

Unloads the FMU and all its instances and frees the allocated memory.

Arguments

fmu::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

fmu::FMU3: Mutable struct representing a FMU and all it instantiated instances in the FMI 3.0 Standard.

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.22]: 2.1.4 Inquire Platform and Version Number of Header Files

FMISpec2.0.2[p.16]: 2.1.2 Platform Dependent Definitions

See also fmi2Unload, fmi3Unload.

FMI.fmiReload — Function

fmiReload(fmu::Union{FMU2, FMU3})

Reloads the FMU-binary. This is useful, if the FMU does not support a clean reset implementation.

Arguments

fmu::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

fmu::FMU3: Mutable struct representing an FMU in the FMI 3.0 Standard.

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.22]: 2.1.4 Inquire Platform and Version Number of Header Files

FMISpec2.0.2[p.16]: 2.1.2 Platform Dependent Definitions

See also fmi2Reload, fmi3Reload.

Conversion functions

FMI.fmiStringToValueReference — Function

fmiStringToValueReference(dataStruct::Union{FMU2, fmi2ModelDescription, FMU3, fmmi3ModelDescription}, identifier::Union{String, AbstractArray{String}})

Returns the ValueReference coresponding to the variable identifier.

Arguments

dataStruct::Union{FMU2, fmi2ModelDescription, FMU3, fmmi3ModelDescription}: Model of the type FMU2/FMU3 or the Model Description of fmi2/fmi3. Same for Model of type FMU3 or the Model Description of fmi3

identifier::Union{String, AbstractArray{String}}: Variable identifier in type String or as a 1-dimensional AbstractArray containing elements of type String

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.22]: 2.1.4 Inquire Platform and Version Number of Header Files

FMISpec2.0.2[p.16]: 2.1.2 Platform Dependent Definitions

See also fmi2StringToValueReference, fmi3StringToValueReference.

External/additional functions

FMI.fmiGetDependencies — Function

fmiGetDependencies(fmu::FMU2)

Building dependency matrix dim x dim for fast look-ups on variable dependencies (dim is number of states).

Arguments

fmu::FMU2: Mutable Struct representing a FMU.

Retruns

fmu.dependencies::Matrix{Union{fmi2DependencyKind, Nothing}}: Returns the FMU's dependency-matrix for fast look-ups on dependencies between value references. Entries are from type fmi2DependencyKind.

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.22]: 2.1.4 Inquire Platform and Version Number of Header Files

FMISpec2.0.2[p.16]: 2.1.2 Platform Dependent Definitions

See also fmi2GetDependencies.

FMI.fmiInfo — Function

fmiInfo(str::fmi2Struct)

Prints FMU-specific information into the REPL.

Arguments

str::Union{fmi2Struct, fmi3Struct}: Representative for an FMU in the FMI 2.0.2 Standard or FMI 3.0 Standard. Other notation:

More detailed: fmi2Struct = Union{FMU2, FMU2Component} More detailed: fmi3StructMD = Union{FMU3, FMU3Instance, fmi3ModelDescription}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

str::FMU3: Mutable struct representing an FMU in the FMI 3.0 Standard.

str::FMU3Instance: Mutable struct represents a pointer to an FMU specific data structure that contains the information needed. Also in FMI 3.0 Standard.

Returns

Prints FMU related information.

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec3.0 Link: https://fmi-standard.org/

See also fmi2Info, fmi2Struct, FMU2, FMU2Component, fmi3Info, fmi3Struct, FMU3, FMU3Instance.

Visualize simulation results

Missing docstring.

Missing docstring for fmiPlot. Check Documenter's build log for details.