Version independent FMU functions

Parsing variable names to ValueReferences

Missing docstring.

Missing docstring for fmiString2ValueReference. Check Documenter's build log for details.

Opening and closing FMUs

Missing docstring.

Missing docstring for fmiUnzip. Check Documenter's build log for details.

FMI.fmiLoad — Function

Load FMUs independent of the FMI version, currently supporting version 2.0.X.

FMI.fmiUnload — Function

Unloads the FMU and all its instances and frees the allocated memory.

Reading the model description

FMI.fmiGetModelName — Function

Returns the tag 'modelName' from the model description.

FMI.fmiGetGUID — Function

Returns the tag 'guid' from the model description.

FMI.fmiGetGenerationTool — Function

Returns the tag 'generationtool' from the model description.

FMI.fmiGetGenerationDateAndTime — Function

Returns the tag 'generationdateandtime' from the model description.

FMI.fmiGetVariableNamingConvention — Function

Returns the tag 'varaiblenamingconvention' from the model description.

FMI.fmiGetNumberOfEventIndicators — Function

Returns the tag 'numberOfEventIndicators' from the model description.

FMI.fmiCanGetSetState — Function

Returns true, if the FMU supports the getting/setting of states

FMI.fmiCanSerializeFMUstate — Function

Returns true, if the FMU state can be serialized

FMI.fmiProvidesDirectionalDerivative — Function

Returns true, if the FMU provides directional derivatives

FMI.fmiIsCoSimulation — Function

Returns true, if the FMU supports co simulation

FMI.fmiIsModelExchange — Function

Returns true, if the FMU supports model exchange

FMI.fmiInfo — Function

Prints FMU-specific information into the REPL.

Simulate FMU

FMI.fmiSimulate — Function

Simulate an fmu according to its standard from 0.0 to t_stop.

FMI.fmiSimulateCS — Function

Simulate an CoSimulation fmu according to its standard from 0.0 to t_stop.

FMI.fmiSimulateME — Function

Simulate an ModelExchange fmu according to its standard from 0.0 to t_stop.

Visualize simulation results

Missing docstring.

Missing docstring for fmiPlot. Check Documenter's build log for details.