Version independent library functions

Creation, Destruction FMU

FMI.fmiInstantiate! — Function

fmiInstantiate!(fmu::FMU2; pushComponents::Bool = true, visible::Bool = false, loggingOn::Bool = false, externalCallbacks::Bool = false,
                      logStatusOK::Bool=true, logStatusWarning::Bool=true, logStatusDiscard::Bool=true, logStatusError::Bool=true, logStatusFatal::Bool=true, logStatusPending::Bool=true)

Creates a new instance of the FMU, version independent.

Create a new instance of the given fmu, adds a logger if logginOn == true.

Arguments

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

Keywords

pushComponents::Bool = true: pushComponents if the item component should be inserted in fmu.components(default = true).

visible::Bool = false: visible if the FMU should be started with graphic interface, if supported (default=false)

loggingOn::Bool = false: loggingOn if the FMU should log and display function calls (default=false)

externalCallbacks::Bool = false: externalCallbacks if an external DLL should be used for the fmi2CallbackFunctions, this may improve readability of logging messages (default=false)

logStatusOK::Bool=true: logStatusOK whether to log status of kind fmi2OK (default=true)

logStatusWarning::Bool=true: logStatusWarning whether to log status of kind fmi2Warning (default=true)

logStatusDiscard::Bool=true: logStatusDiscard whether to log status of kind fmi2Discard (default=true)

logStatusError::Bool=true: logStatusError whether to log status of kind fmi2Error (default=true)

logStatusFatal::Bool=true: logStatusFatal whether to log status of kind fmi2Fatal (default=true)

logStatusPending::Bool=true: logStatusPending whether to log status of kind fmi2Pending (default=true)

Returns

nothing: if the instantiation failed. In addition, an error message appears.

component: Returns the instance of a new FMU component.

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.19]: 2.1.5 Creation, Destruction and Logging of FMU Instances

See also fmi2Instantiate!, fmi2Instantiate, FMU2.

FMI.fmiFreeInstance! — Function

fmiFreeInstance!(str::fmi2Struct)

Frees the allocated memory of the last instance of the FMU.

Arguments

str::fmi2Struct: Representative for an FMU in the FMI 2.0.2 Standard.

More detailed: fmi2Struct = Union{FMU2, FMU2Component}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

Returns

status::fmi2Status: returned by all functions to indicate the success of the function call

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.23]: 2.1.6 Initialization, Termination, and Resetting an FMU

FMISpec2.0.2[p.18]: 2.1.3 Status Returned by Functions

See also fmi2FreeInstance, fmi2Struct, FMU2, FMU2Component.

Platform and Version number

FMI.fmiGetTypesPlatform — Function

fmiGetTypesPlatform(str::fmi2Struct)

Returns the header file used to compile the FMU. By default returns default, version independent.

Arguments

str::fmi2Struct: Representative for an FMU in the FMI 2.0.2 Standard.

More detailed: fmi2Struct = Union{FMU2, FMU2Component}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

Returns

See also fmi2GetVersion, fmi2Struct, FMU2, FMU2Component.

FMI.fmiGetVersion — Function

fmiGetVersion(str::fmi2Struct)

Returns the version of the FMU, version independent.

Arguments

str::fmi2Struct: Representative for an FMU in the FMI 2.0.2 Standard.

More detailed: fmi2Struct = Union{FMU2, FMU2Component}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

Returns

Returns a string from the address of a C-style (NUL-terminated) string. The string represents the version of the “fmi2Functions.h” header file which was used to compile the functions of the FMU. The function returns “fmiVersion” which is defined in this header file. The standard header file as documented in this specification has version “2.0”

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.22]: 2.1.4 Inquire Platform and Version Number of Header Files

FMISpec2.0.2[p.16]: 2.1.2 Platform Dependent Definitions

See also fmi2GetVersion, unsafe_string, fmi2Struct, FMU2, FMU2Component.

FMI.fmiSetDebugLogging — Function

fmiSetDebugLogging(str::fmi2Struct)

Control the use of the logging callback function, version independent.

Arguments

str::fmi2Struct: Representative for an FMU in the FMI 2.0.2 Standard.

More detailed: fmi2Struct = Union{FMU2, FMU2Component}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.22]: 2.1.5 Creation, Destruction and Logging of FMU Instances

Initialization, Termination and Destruction

FMI.fmiSetupExperiment — Function

fmiSetupExperiment(str::fmi2Struct, c::FMU2Component, startTime::Union{Real, Nothing} = nothing, stopTime::Union{Real, Nothing} = nothing; tolerance::Union{Real, Nothing} = nothing)

Initialize the Simulation boundries

Arguments

str::fmi2Struct: Representative for an FMU in the FMI 2.0.2 Standard.

More detailed: fmi2Struct = Union{FMU2, FMU2Component}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

startTime::Union{Real, Nothing} = nothing: startTime is a real number which sets the value of starting time of the experiment. The default value is set automatically if doing nothing (default = nothing).

stopTime::Union{Real, Nothing} = nothing: stopTime is a real number which sets the value of ending time of the experiment. The default value is set automatically if doing nothing (default = nothing).

Keywords

tolerance::Union{Real, Nothing} = nothing: tolerance is a real number which sets the value of tolerance range. The default value is set automatically if doing nothing (default = nothing).

Returns

Returns a warning if str.state is not called in fmi2ComponentStateInstantiated.

status::fmi2Status: returned by all functions to indicate the success of the function call

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.23]: 2.1.6 Initialization, Termination, and Resetting an FMU

FMISpec2.0.2[p.18]: 2.1.3 Status Returned by Functions

See also fmi2SetupExperiment, fmi2Struct, FMU2, FMU2Component.

FMI.fmiEnterInitializationMode — Function

fmiEnterInitializationMode(str::fmi2Struct)

Informs the FMU to enter initializaton mode, version independent.

Arguments

str::fmi2Struct: Representative for an FMU in the FMI 2.0.2 Standard.

More detailed: fmi2Struct = Union{FMU2, FMU2Component}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

Returns

Returns a warning if str.state is not called in fmi2ComponentStateInstantiated.

status::fmi2Status: returned by all functions to indicate the success of the function call

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.23]: 2.1.6 Initialization, Termination, and Resetting an FMU

FMISpec2.0.2[p.18]: 2.1.3 Status Returned by Functions

See also fmi2EnterInitializationMode, fmi2Struct, FMU2, FMU2Component.

FMI.fmiExitInitializationMode — Function

fmiExitInitializationMode(str::fmi2Struct)

Informs the FMU to exit initialization mode, version independent.

Arguments

str::fmi2Struct: Representative for an FMU in the FMI 2.0.2 Standard.

More detailed: fmi2Struct = Union{FMU2, FMU2Component}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

Returns

Returns a warning if str.state is not called in fmi2ComponentStateInitializationMode.

status::fmi2Status: returned by all functions to indicate the success of the function call

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.23]: 2.1.6 Initialization, Termination, and Resetting an FMU

FMISpec2.0.2[p.18]: 2.1.3 Status Returned by Functions

See also fmi2ExitInitializationMode

FMI.fmiTerminate — Function

Informs the FMU that the simulation run is terminated, version independent.

FMI.fmiReset — Function

Resets the FMU after a simulation run, version independent.

Get/Set variable values

FMI.fmiGetReal — Function

Returns the real values of an array of variables

FMI.fmiGetReal! — Function

Writes the real values of an array of variables in the given field

FMI.fmiSetReal — Function

Set the values of an array of real variables

FMI.fmiGetInteger — Function

Returns the integer values of an array of variables

FMI.fmiGetInteger! — Function

Writes the integer values of an array of variables in the given field

FMI.fmiSetInteger — Function

Set the values of an array of integer variables

FMI.fmiGetBoolean — Function

Returns the boolean values of an array of variables

FMI.fmiGetBoolean! — Function

Writes the boolean values of an array of variables in the given field

FMI.fmiSetBoolean — Function

Set the values of an array of boolean variables

FMI.fmiGetString — Function

Returns the string values of an array of variables

FMI.fmiGetString! — Function

Writes the string values of an array of variables in the given field

FMI.fmiSetString — Function

Set the values of an array of string variables

FMU state Functions

FMI.fmiGetFMUstate — Function

Returns the FMU state of the fmu

FMI.fmiSetFMUstate — Function

Sets the FMU to the given state

FMI.fmiFreeFMUstate! — Function

Free the memory for the allocated FMU state

FMI.fmiSerializedFMUstateSize — Function

Returns the size of the byte vector the FMU can be stored in

FMI.fmiSerializeFMUstate — Function

Serialize the data in the FMU state pointer

FMI.fmiDeSerializeFMUstate — Function

Deserialize the data in the FMU state pointer

Partial Derivatives

FMI.fmiGetDirectionalDerivative — Method

Returns the values of the directional derivatives.

CoSimulation specific Functions

FMI.fmiDoStep — Method

Does one step in the CoSimulation FMU

ModelExchange specific Functions

FMI.fmiSetTime — Function

Set a time instant

FMI.fmiSetContinuousStates — Function

Set a new (continuous) state vector

FMI.fmiNewDiscreteStates — Function

Returns the next discrete states

FMI.fmiEnterContinuousTimeMode — Function

The model enters Continuous-Time Mode

FMI.fmiCompletedIntegratorStep — Function

This function must be called by the environment after every completed step

FMI.fmiGetDerivatives — Function

Compute state derivatives at the current time instant and for the current states

FMI.fmiGetEventIndicators — Function

Returns the event indicators of the FMU

FMI.fmiGetContinuousStates — Function

Return the new (continuous) state vector x

FMI.fmiGetNominalsOfContinuousStates — Function

Return the new (continuous) state vector x