Version independent library functions

Creation, Destruction FMU

FMI.fmiInstantiate! — Function

Creates a new instance of the FMU, version independent.

FMI.fmiFreeInstance! — Function

Frees the allocated memory of the last instance of the FMU.

Platform and Version number

FMI.fmiGetTypesPlatform — Function

Returns the header file used to compile the FMU. By default returns default, version independent.

FMI.fmiGetVersion — Function

Returns the version of the FMU, version independent.

FMI.fmiSetDebugLogging — Function

Control the use of the logging callback function, version independent.

Initialization, Termination and Destruction

FMI.fmiSetupExperiment — Function

Initialize the Simulation boundries

FMI.fmiEnterInitializationMode — Function

Informs the FMU to enter initializaton mode, version independent.

FMI.fmiExitInitializationMode — Function

Informs the FMU to exit initialization mode, version independent.

FMI.fmiTerminate — Function

Informs the FMU that the simulation run is terminated, version independent.

FMI.fmiReset — Function

Resets the FMU after a simulation run, version independent.

Get/Set variable values

FMI.fmiGetReal — Function

Returns the real values of an array of variables

FMI.fmiGetReal! — Function

Writes the real values of an array of variables in the given field

FMI.fmiSetReal — Function

Set the values of an array of real variables

FMI.fmiGetInteger — Function

Returns the integer values of an array of variables

FMI.fmiGetInteger! — Function

Writes the integer values of an array of variables in the given field

FMI.fmiSetInteger — Function

Set the values of an array of integer variables

FMI.fmiGetBoolean — Function

Returns the boolean values of an array of variables

FMI.fmiGetBoolean! — Function

Writes the boolean values of an array of variables in the given field

FMI.fmiSetBoolean — Function

Set the values of an array of boolean variables

FMI.fmiGetString — Function

Returns the string values of an array of variables

FMI.fmiGetString! — Function

Writes the string values of an array of variables in the given field

FMI.fmiSetString — Function

Set the values of an array of string variables

FMU state Functions

FMI.fmiGetFMUstate — Function

Returns the FMU state of the fmu

FMI.fmiSetFMUstate — Function

Sets the FMU to the given state

Missing docstring.

Missing docstring for fmiFreeFMUstate. Check Documenter's build log for details.

FMI.fmiSerializedFMUstateSize — Function

Returns the size of the byte vector the FMU can be stored in

FMI.fmiSerializeFMUstate — Function

Serialize the data in the FMU state pointer

FMI.fmiDeSerializeFMUstate — Function

Deserialize the data in the FMU state pointer

Partial Derivatives

FMI.fmiGetDirectionalDerivative — Method

Returns the values of the directional derivatives.

CoSimulation specific Functions

FMI.fmiDoStep — Method

Does one step in the CoSimulation FMU

ModelExchange specific Functions

FMI.fmiSetTime — Function

Set a time instant

FMI.fmiSetContinuousStates — Function

Set a new (continuous) state vector

FMI.fmiNewDiscreteStates — Function

Returns the next discrete states

FMI.fmiEnterContinuousTimeMode — Function

The model enters Continuous-Time Mode

FMI.fmiCompletedIntegratorStep — Function

This function must be called by the environment after every completed step

FMI.fmiGetDerivatives — Function

Compute state derivatives at the current time instant and for the current states

FMI.fmiGetEventIndicators — Function

Returns the event indicators of the FMU

FMI.fmiGetContinuousStates — Function

Return the new (continuous) state vector x

FMI.fmiGetNominalsOfContinuousStates — Function

Return the new (continuous) state vector x