Functions in FMI Import/Core .jl

FMICore.logInfo — Function

Prints a message with level info if the log level allows it.

FMICore.logWarning — Function

Prints a message with level warn if the log level allows it.

FMICore.logError — Function

Prints a message with level error if the log level allows it.

FMIImport.loadBinary — Function

loadBinary(fmu::FMU2)

load pointers to fmus c functions from shared library handle (provided byfmu.libHandle`)

loadBinary(fmu::FMU3)

load pointers to fmus c functions from shared library handle (provided byfmu.libHandle`)

FMICore.eval! — Function

ToDo: Doc String