Library functions

Creation, Destruction FMU

FMI.fmi2Instantiate — Function

Source: FMISpec2.0.2[p.19]: 2.1.5 Creation, Destruction and Logging of FMU Instances

The function returns a new instance of an FMU.

FMI.fmi2FreeInstance — Function

Source: FMISpec2.0.2[p.22]: 2.1.5 Creation, Destruction and Logging of FMU Instances

Disposes the given instance, unloads the loaded model, and frees all the allocated memory and other resources that have been allocated by the functions of the FMU interface. If a null pointer is provided for “c”, the function call is ignored (does not have an effect).

Platform and Version number

FMI.fmi2GetVersion — Function

Source: FMISpec2.0.2[p.22]: 2.1.4 Inquire Platform and Version Number of Header Files

Returns the version of the “fmi2Functions.h” header file which was used to compile the functions of the FMU. The function returns “fmiVersion” which is defined in this header file. The standard header file as documented in this specification has version “2.0”

Returns the version of the FMI Standard used in this FMU

For more information call ?fmi2GetVersion

FMI.fmi2GetTypesPlatform — Function

Source: FMISpec2.0.2[p.22]: 2.1.4 Inquire Platform and Version Number of Header Files

Returns the string to uniquely identify the “fmi2TypesPlatform.h” header file used for compilation of the functions of the FMU. The standard header file, as documented in this specification, has fmi2TypesPlatform set to “default” (so this function usually returns “default”).

Returns a string representing the header file used to compile the fmi2 functions.function

Returns "default" by default

For more information call ?fmi2GetTypesPlatform

FMI.fmi2SetDebugLogging — Function

Source: FMISpec2.0.2[p.22]: 2.1.5 Creation, Destruction and Logging of FMU Instances

The function controls debug logging that is output via the logger function callback. If loggingOn = fmi2True, debug logging is enabled, otherwise it is switched off.

Set the DebugLogger for the FMU

For more information call ?fmi2SetDebugLogging

Set the DebugLogger for the FMU

For more information call ?fmi2SetDebugLogging

Initialization, Termination and Destruction

FMI.fmi2SetupExperiment — Function

Source: FMISpec2.0.2[p.23]: 2.1.6 Initialization, Termination, and Resetting an FMU

Informs the FMU to setup the experiment. This function must be called after fmi2Instantiate and before fmi2EnterInitializationMode is called.The function controls debug logging that is output via the logger function callback. If loggingOn = fmi2True, debug logging is enabled, otherwise it is switched off.

Set the start and end time for a simulation, can only be called after fmi2Instantiate and before fmi2EnterInitializationMode

For more information call ?fmi2SetupExperiment

Setup the simulation but without defining all of the parameters

For more information call ?fmi2SetupExperiment

Set the start and end time for a simulation, can only be called after fmi2Instantiate and before fmi2EnterInitializationMode

For more information call ?fmi2SetupExperiment

Setup the simulation but without defining all of the parameters

For more information call ?fmi2SetupExperiment

FMI.fmi2EnterInitializationMode — Function

Source: FMISpec2.0.2[p.23]: 2.1.6 Initialization, Termination, and Resetting an FMU

Informs the FMU to enter Initialization Mode. Before calling this function, all variables with attribute <ScalarVariable initial = "exact" or "approx"> can be set with the “fmi2SetXXX” functions (the ScalarVariable attributes are defined in the Model Description File, see section 2.2.7). Setting other variables is not allowed. Furthermore, fmi2SetupExperiment must be called at least once before calling fmi2EnterInitializationMode, in order that startTime is defined.

FMU enters Initialization mode

For more information call ?fmi2EnterInitializationMode

FMI.fmi2ExitInitializationMode — Function

Source: FMISpec2.0.2[p.23]: 2.1.6 Initialization, Termination, and Resetting an FMU

Informs the FMU to exit Initialization Mode.

FMU exits Initialization mode

For more information call ?fmi2ExitInitializationMode

FMI.fmi2Terminate — Function

Source: FMISpec2.0.2[p.24]: 2.1.6 Initialization, Termination, and Resetting an FMU

Informs the FMU that the simulation run is terminated.

Informs FMU that simulation run is terminated

For more information call ?fmi2Terminate

FMI.fmi2Reset — Method

Source: FMISpec2.0.2[p.24]: 2.1.6 Initialization, Termination, and Resetting an FMU

Is called by the environment to reset the FMU after a simulation run. The FMU goes into the same state as if fmi2Instantiate would have been called.

Get/Set variable values

FMI.fmi2GetReal — Function

Get the values of an array of fmi2Real variables

For more information call ?fmi2GetReal

Get the value of a fmi2Real variable

For more information call ?fmi2GetReal

Get the values of an array of fmi2Real variables by variable name

For more information call ?fmi2GetReal

Get the value of a fmi2Real variable by variable name

For more information call ?fmi2GetReal

Get the values of an array of fmi2Real variables

For more information call ?fmi2GetReal

Get the value of a fmi2Real variable

For more information call ?fmi2GetReal

Get the values of an array of fmi2Real variables by variable name

For more information call ?fmi2GetReal

Get the value of a fmi2Real variable by variable name

For more information call ?fmi2GetReal

FMI.fmi2GetReal! — Function

Source: FMISpec2.0.2[p.24]: 2.1.7 Getting and Setting Variable Values

Functions to get and set values of variables idetified by their fmi2valueReference

Get the values of an array of fmi2Real variables

For more information call ?fmi2GetReal

Get the values of an array of fmi2Real variables by variable name

For more information call ?fmi2GetReal

Get the values of an array of fmi2Real variables

For more information call ?fmi2GetReal

Get the values of an array of fmi2Real variables by variable name

For more information call ?fmi2GetReal

FMI.fmi2SetReal — Function

Source: FMISpec2.0.2[p.24]: 2.1.7 Getting and Setting Variable Values

Functions to get and set values of variables idetified by their fmi2valueReference

Set the values of an array of fmi2Real variables

For more information call ?fmi2SetReal

Set the value of a fmi2Real variable

For more information call ?fmi2SetReal

Set the values of an array of fmi2Real variables by variable name

For more information call ?fmi2SetReal

Set the value of a fmi2Real variable by variable name

For more information call ?fmi2SetReal

Set the values of an array of fmi2Real variables

For more information call ?fmi2SetReal

Set the value of a fmi2Real variable

For more information call ?fmi2SetReal

Set the values of an array of fmi2Real variables by variable name

For more information call ?fmi2SetReal

Set the value of a fmi2Real variable by variable name

For more information call ?fmi2SetReal

FMI.fmi2GetInteger — Function

Get the values of an array of fmi2Integer variables

For more information call ?fmi2GetInteger

Get the value of a fmi2Integer variable

For more information call ?fmi2GetInteger

Get the values of an array of fmi2Integer variables by variable name

For more information call ?fmi2GetInteger

Get the value of a fmi2Integer variable by variable name

For more information call ?fmi2GetInteger

Get the values of an array of fmi2Integer variables

For more information call ?fmi2GetInteger

Get the value of a fmi2Integer variable

For more information call ?fmi2GetInteger

Get the values of an array of fmi2Integer variables by variable name

For more information call ?fmi2GetInteger

Get the value of a fmi2Integer variable by variable name

For more information call ?fmi2GetInteger

FMI.fmi2GetInteger! — Function

Source: FMISpec2.0.2[p.24]: 2.1.7 Getting and Setting Variable Values

Functions to get and set values of variables idetified by their fmi2valueReference

Get the values of an array of fmi2Integer variables

For more information call ?fmi2GetInteger

Get the values of an array of fmi2Integer variables by variable name

For more information call ?fmi2GetInteger

Get the values of an array of fmi2Integer variables

For more information call ?fmi2GetInteger

Get the values of an array of fmi2Integer variables by variable name

For more information call ?fmi2GetInteger

FMI.fmi2SetInteger — Function

Source: FMISpec2.0.2[p.24]: 2.1.7 Getting and Setting Variable Values

Functions to get and set values of variables idetified by their fmi2valueReference

Set the values of an array of fmi2Integer variables

For more information call ?fmi2SetInteger

Get the value of a fmi2Integer variable

For more information call ?fmi2SetInteger

Set the values of an array of fmi2Integer variables by variable name

For more information call ?fmi2SetInteger

Set the value of a fmi2Integer variable by variable name

For more information call ?fmi2SetInteger

Set the values of an array of fmi2Integer variables

For more information call ?fmi2SetInteger

Get the value of a fmi2Integer variable

For more information call ?fmi2SetInteger

Set the values of an array of fmi2Integer variables by variable name

For more information call ?fmi2SetInteger

Set the value of a fmi2Integer variable by variable name

For more information call ?fmi2SetInteger

FMI.fmi2GetBoolean — Function

Get the values of an array of fmi2Boolean variables

For more information call ?fmi2GetBoolean

Get the value of a fmi2Boolean variable

For more information call ?fmi2GetBoolean

Get the values of an array of fmi2Boolean variables by variable name

For more information call ?fmi2GetBoolean

Get the value of a fmi2Boolean variable by variable name

For more information call ?fmi2GetBoolean

Get the values of an array of fmi2Boolean variables

For more information call ?fmi2GetBoolean

Get the value of a fmi2Boolean variable

For more information call ?fmi2GetBoolean

Get the values of an array of fmi2Boolean variables by variable name

For more information call ?fmi2GetBoolean

Get the value of a fmi2Boolean variable by variable name

For more information call ?fmi2GetBoolean

FMI.fmi2GetBoolean! — Function

Source: FMISpec2.0.2[p.24]: 2.1.7 Getting and Setting Variable Values

Functions to get and set values of variables idetified by their fmi2valueReference

Get the values of an array of fmi2Boolean variables

For more information call ?fmi2GetBoolean

Get the values of an array of fmi2Boolean variables by variable name

For more information call ?fmi2GetBoolean

Get the values of an array of fmi2Boolean variables

For more information call ?fmi2GetBoolean

Get the values of an array of fmi2Boolean variables by variable name

For more information call ?fmi2GetBoolean

FMI.fmi2SetBoolean — Function

Source: FMISpec2.0.2[p.24]: 2.1.7 Getting and Setting Variable Values

Functions to get and set values of variables idetified by their fmi2valueReference

Set the values of an array of fmi2Boolean variables

For more information call ?fmi2SetBoolean

Set the value of a fmi2Boolean variable

For more information call ?fmi2SetBoolean

Set the values of an array of fmi2Boolean variables by variable name

For more information call ?fmi2SetBoolean

Set the value of a fmi2Boolean variable by variable name

For more information call ?fmi2SetBoolean

Set the values of an array of fmi2Boolean variables

For more information call ?fmi2SetBoolean

Set the value of a fmi2Boolean variable

For more information call ?fmi2SetBoolean

Set the values of an array of fmi2Boolean variables by variable name

For more information call ?fmi2SetBoolean

Set the value of a fmi2Boolean variable by variable name

For more information call ?fmi2SetBoolean

FMI.fmi2GetString — Function

Get the values of an array of fmi2String variables

For more information call ?fmi2GetString

Get the value of a fmi2String variable

For more information call ?fmi2GetString

Get the values of an array of fmi2String variables by variable name

For more information call ?fmi2GetString

Get the value of a fmi2String variable by variable name

For more information call ?fmi2GetString

Get the values of an array of fmi2String variables

For more information call ?fmi2GetString

Get the value of a fmi2String variable

For more information call ?fmi2GetString

Get the values of an array of fmi2String variables by variable name

For more information call ?fmi2GetString

Get the value of a fmi2String variable by variable name

For more information call ?fmi2GetString

FMI.fmi2GetString! — Function

Source: FMISpec2.0.2[p.24]: 2.1.7 Getting and Setting Variable Values

Functions to get and set values of variables idetified by their fmi2valueReference

Get the values of an array of fmi2String variables

For more information call ?fmi2GetString

Get the values of an array of fmi2String variables by variable name

For more information call ?fmi2GetString

Get the values of an array of fmi2String variables

For more information call ?fmi2GetString

Get the values of an array of fmi2String variables by variable name

For more information call ?fmi2GetString

FMI.fmi2SetString — Function

Source: FMISpec2.0.2[p.24]: 2.1.7 Getting and Setting Variable Values

Functions to get and set values of variables idetified by their fmi2valueReference

Set the values of an array of fmi2String variables

For more information call ?fmi2SetString

Set the values of a fmi2String variable

For more information call ?fmi2SetString

Set the values of an array of fmi2String variables by variable name

For more information call ?fmi2SetString

Set the value of a fmi2String variable by variable name

For more information call ?fmi2SetString

Set the values of an array of fmi2String variables

For more information call ?fmi2SetString

Set the values of a fmi2String variable

For more information call ?fmi2SetString

Set the values of an array of fmi2String variables by variable name

For more information call ?fmi2SetString

Set the value of a fmi2String variable by variable name

For more information call ?fmi2SetString

FMU state Functions

FMI.fmi2GetFMUstate — Function

Source: FMISpec2.0.2[p.26]: 2.1.8 Getting and Setting the Complete FMU State

fmi2GetFMUstate makes a copy of the internal FMU state and returns a pointer to this copy

Get the pointer to the current FMU state

For more information call ?fmi2GetFMUstate

Get the pointer to the current FMU state

For more information call ?fmi2GetFMUstate

FMI.fmi2SetFMUstate — Function

Source: FMISpec2.0.2[p.26]: 2.1.8 Getting and Setting the Complete FMU State

fmi2SetFMUstate copies the content of the previously copied FMUstate back and uses it as actual new FMU state.

Set the FMU state to the given fmi2FMUstate

For more information call ?fmi2SetFMUstate

FMI.fmi2FreeFMUstate — Function

Source: FMISpec2.0.2[p.26]: 2.1.8 Getting and Setting the Complete FMU State

fmi2FreeFMUstate frees all memory and other resources allocated with the fmi2GetFMUstate call for this FMUstate.

Free the allocated memory for the FMU state

For more information call ?fmi2FreeFMUstate

Free the allocated memory for the FMU state

For more information call ?fmi2FreeFMUstate

FMI.fmi2SerializedFMUstateSize — Function

Source: FMISpec2.0.2[p.26]: 2.1.8 Getting and Setting the Complete FMU State

fmi2SerializedFMUstateSize returns the size of the byte vector, in order that FMUstate can be stored in it.

Returns the size of a byte vector the FMU can be stored in

For more information call ?fmi2SerzializedFMUstateSize

Returns the size of a byte vector the FMU can be stored in

For more information call ?fmi2SerzializedFMUstateSize

FMI.fmi2SerializeFMUstate — Function

Source: FMISpec2.0.2[p.26]: 2.1.8 Getting and Setting the Complete FMU State

fmi2SerializeFMUstate serializes the data which is referenced by pointer FMUstate and copies this data in to the byte vector serializedState of length size,

Serialize the data in the FMU state pointer

For more information call ?fmi2SerzializeFMUstate

Serialize the data in the FMU state pointer

For more information call ?fmi2SerzializeFMUstate

FMI.fmi2DeSerializeFMUstate — Function

Source: FMISpec2.0.2[p.26]: 2.1.8 Getting and Setting the Complete FMU State

fmi2DeSerializeFMUstate deserializes the byte vector serializedState of length size, constructs a copy of the FMU state and returns FMUstate, the pointer to this copy.

Deserialize the data in the serializedState fmi2Byte field

For more information call ?fmi2DeSerzializeFMUstate

Deserialize the data in the serializedState fmi2Byte field

For more information call ?fmi2DeSerzializeFMUstate

Partial Derivatives

FMI.fmi2GetDirectionalDerivative! — Function

Source: FMISpec2.0.2[p.26]: 2.1.9 Getting Partial Derivatives

This function computes the directional derivatives of an FMU.

FMI.fmi2GetDirectionalDerivative — Function

Computes directional derivatives

For more information call ?fmi2GetDirectionalDerivatives

Computes directional derivatives

For more information call ?fmi2GetDirectionalDerivatives

Computes directional derivatives

For more information call ?fmi2GetDirectionalDerivatives

Computes directional derivatives

For more information call ?fmi2GetDirectionalDerivatives

CoSimulation specific Functions

FMI.fmi2SetRealInputDerivatives — Function

Source: FMISpec2.0.2[p.104]: 4.2.1 Transfer of Input / Output Values and Parameters

Sets the n-th time derivative of real input variables.

Sets the n-th time derivative of real input variables.

For more information call ?fmi2SetRealInputDerivatives

FMI.fmi2GetRealOutputDerivatives — Function

Source: FMISpec2.0.2[p.104]: 4.2.1 Transfer of Input / Output Values and Parameters

Retrieves the n-th derivative of output values.

Retrieves the n-th derivative of output values.

For more information call ?fmi2GetRealOutputDerivatives

FMI.fmi2DoStep — Function

Source: FMISpec2.0.2[p.104]: 4.2.2 Computation

The computation of a time step is started.

The computation of a time step is started.

For more information call ?fmi2DoStep

The computation of a time step is started.

For more information call ?fmi2DoStep

The computation of a time step is started.

For more information call ?fmi2DoStep

FMI.fmi2CancelStep — Function

Source: FMISpec2.0.2[p.105]: 4.2.2 Computation

Can be called if fmi2DoStep returned fmi2Pending in order to stop the current asynchronous execution.

Cancels a DoStep that didn't finish correctly

For more information call ?fmi2CancelStep

FMI.fmi2GetStatus — Function

Source: FMISpec2.0.2[p.106]: 4.2.3 Retrieving Status Information from the Slave

Informs the master about the actual status of the simulation run. Which status information is to be returned is specified by the argument fmi2StatusKind.

Informs Master of fmi2Status

For more information call ?fmi2GetStatus

FMI.fmi2GetRealStatus — Function

Source: FMISpec2.0.2[p.106]: 4.2.3 Retrieving Status Information from the Slave

Informs the master about the actual status of the simulation run. Which status information is to be returned is specified by the argument fmi2StatusKind.

Informs Master of fmi2Status of fmi2Real

For more information call ?fmi2GetRealStatus

FMI.fmi2GetIntegerStatus — Function

Source: FMISpec2.0.2[p.106]: 4.2.3 Retrieving Status Information from the Slave

Informs the master about the actual status of the simulation run. Which status information is to be returned is specified by the argument fmi2StatusKind.

Informs Master of fmi2Status of fmi2Integer

For more information call ?fmi2GetIntegerStatus

FMI.fmi2GetBooleanStatus — Function

Source: FMISpec2.0.2[p.106]: 4.2.3 Retrieving Status Information from the Slave

Informs the master about the actual status of the simulation run. Which status information is to be returned is specified by the argument fmi2StatusKind.

Informs Master of fmi2Status of fmi2Boolean

For more information call ?fmi2GetBooleanStatus

FMI.fmi2GetStringStatus — Function

Source: FMISpec2.0.2[p.106]: 4.2.3 Retrieving Status Information from the Slave

Informs the master about the actual status of the simulation run. Which status information is to be returned is specified by the argument fmi2StatusKind.

Informs Master of fmi2Status of fmi2String

For more information call ?fmi2GetStringStatus

ModelExchange specific Functions

FMI.fmi2SetTime — Function

Source: FMISpec2.0.2[p.83]: 3.2.1 Providing Independent Variables and Re-initialization of Caching

Set a new time instant and re-initialize caching of variables that depend on time, provided the newly provided time value is different to the previously set time value (variables that depend solely on constants or parameters need not to be newly computed in the sequel, but the previously computed values can be reused).

Set a new time instant

For more information call ?fmi2SetTime

Set a new time instant

For more information call ?fmi2SetTime

FMI.fmi2SetContinuousStates — Function

Source: FMISpec2.0.2[p.83]: 3.2.1 Providing Independent Variables and Re-initialization of Caching

Set a new (continuous) state vector and re-initialize caching of variables that depend on the states. Argument nx is the length of vector x and is provided for checking purposes

Set a new (continuous) state vector

For more information call ?fmi2SetContinuousStates

Set a new (continuous) state vector

For more information call ?fmi2SetContinuousStates

FMI.fmi2EnterEventMode — Function

Source: FMISpec2.0.2[p.84]: 3.2.2 Evaluation of Model Equations

The model enters Event Mode from the Continuous-Time Mode and discrete-time equations may become active (and relations are not “frozen”).

The model enters Event Mode

For more information call ?fmi2EnterEventMode

The model enters Event Mode

FMI.fmi2NewDiscreteStates — Function

Source: FMISpec2.0.2[p.84]: 3.2.2 Evaluation of Model Equations

The FMU is in Event Mode and the super dense time is incremented by this call.

Returns the next discrete states

For more information call ?fmi2NewDiscreteStates

Returns the next discrete states

For more information call ?fmi2NewDiscretestates

FMI.fmi2EnterContinuousTimeMode — Function

Source: FMISpec2.0.2[p.85]: 3.2.2 Evaluation of Model Equations

The model enters Continuous-Time Mode and all discrete-time equations become inactive and all relations are “frozen”. This function has to be called when changing from Event Mode (after the global event iteration in Event Mode over all involved FMUs and other models has converged) into Continuous-Time Mode.

The model enters Continuous-Time Mode

For more information call ?fmi2EnterContinuousTimeMode

FMI.fmi2CompletedIntegratorStep! — Function

Source: FMISpec2.0.2[p.85]: 3.2.2 Evaluation of Model Equations

This function must be called by the environment after every completed step of the integrator provided the capability flag completedIntegratorStepNotNeeded = false.

FMI.fmi2GetDerivatives — Function

Source: FMISpec2.0.2[p.86]: 3.2.2 Evaluation of Model Equations

Compute state derivatives at the current time instant and for the current states.

Compute state derivatives at the current time instant and for the current states.

For more information call ?fmi2GetDerivatives

Compute state derivatives at the current time instant and for the current states.

For more information call ?fmi2GetDerivatives

FMI.fmi2GetEventIndicators — Function

Source: FMISpec2.0.2[p.86]: 3.2.2 Evaluation of Model Equations

Compute event indicators at the current time instant and for the current states.

Returns the event indicators of the FMU

For more information call ?fmi2GetEventIndicators

Returns the event indicators of the FMU

For more information call ?fmi2GetEventIndicators

FMI.fmi2GetContinuousStates — Function

Source: FMISpec2.0.2[p.86]: 3.2.2 Evaluation of Model Equations

Return the new (continuous) state vector x.

Return the new (continuous) state vector x

For more information call ?fmi2GetContinuousStates

Return the new (continuous) state vector x

For more information call ?fmi2GetContinuousStates

FMI.fmi2GetNominalsOfContinuousStates — Function

Source: FMISpec2.0.2[p.86]: 3.2.2 Evaluation of Model Equations

Return the nominal values of the continuous states.

Return the new (continuous) state vector x

For more information call ?fmi2GetNominalsOfContinuousStates

Return the new (continuous) state vector x

For more information call ?fmi2GetNominalsOfContinuousStates