FMI.jl Library Functions

Many of the functions in this library are based on already defined functions of the FMIImport.jl library.

Simulate FMUs

FMI.fmiLoad — Function

fmiLoad(pathToFMU::String; unpackPath=nothing, type=nothing)

Load FMUs independent of the FMI version, currently supporting version 2.0.X and 3.0.

Arguments

pathToFMU::String: String that contains the paths of ziped and unziped FMU folders.

Keywords

unpackPath=nothing: Via optional argument unpackPath, a path to unpack the FMU can be specified (default: system temporary directory).

type::Union{CS, ME, SE} = nothing: Via type, a FMU type can be selected. If none of the unified type set is used, the default value type = nothing will be used.

Returns

Returns the instance of the FMU struct.

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.22]: 2.1.4 Inquire Platform and Version Number of Header Files

FMISpec2.0.2[p.16]: 2.1.2 Platform Dependent Definitions

See also fmi2Load, fmi3Load.

FMI.fmiSimulate — Function

fmiSimulate(str::fmi2Struct, args...; kwargs...)

Start a simulation of the FMU instance str for the matching FMU version and type.

Wrapper for fmi2Simulate.

fmiSimulate(str::fmi3Struct, args...; kwargs...)

Start a simulation of the FMU instance str for the matching FMU version and type.

Wrapper for fmi3Simulate.

FMI.fmiSimulateCS — Function

fmiSimulateCS(str::fmi2Struct, tspan::Union{Tuple{Float64, Float64}, Nothing}=nothing, args...; kwargs...)

Start a simulation of the Co-Simulation FMU instance str.

Wrapper for fmi2SimulateCS.

fmiSimulateCS(str::fmi3Struct, args...; kwargs...)

Start a simulation of the Co-Simulation FMU instance str.

Wrapper for fmi3SimulateCS.

FMI.fmiSimulateME — Function

fmiSimulateME(str::Union{fmi2Struct,fmi3Struct}, tspan::Union{Tuple{Float64, Float64}, Nothing}=nothing;
                tolerance::Union{Real, Nothing} = nothing,
                dt::Union{Real, Nothing} = nothing,
                solver = nothing,
                customFx = nothing,
                recordValues::fmi2ValueReferenceFormat = nothing,
                saveat = nothing,
                x0::Union{AbstractArray{<:Real}, Nothing} = nothing,
                setup::Union{Bool, Nothing} = nothing,
                reset::Union{Bool, Nothing} = nothing,
                instantiate::Union{Bool, Nothing} = nothing,
                freeInstance::Union{Bool, Nothing} = nothing,
                terminate::Union{Bool, Nothing} = nothing,
                inputValueReferences::fmi2ValueReferenceFormat = nothing,
                inputFunction = nothing,
                parameters::Union{Dict{<:Any, <:Any}, Nothing} = nothing,
                dtmax::Union{Real, Nothing} = nothing,
                callbacks = [],
                showProgress::Bool = true,
                kwargs...)

Simulates a FMU instance for the given simulation time interval.

Arguments

str::Union{fmi2StructMD, fmi3StructMD}: Representative for an FMU in the FMI 2.0.2 Standard or FMI 3.0 Standard. Other notation:

More detailed: fmi2Struct = Union{FMU2, FMU2Component} More detailed: fmi3Struct = Union{FMU3, FMU3Instance}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

str::FMU3: Mutable struct representing an FMU in the FMI 3.0 Standard.

str::FMU3Instance: Mutable struct represents a pointer to an FMU specific data structure that contains the information needed. Also in FMI 3.0 Standard.

tspan::Union{Tuple{Float64, Float64}, Nothing}=nothing: Sets the time span as a tuple or the default value from the model description is used.

Keywords

tolerance::Union{Real, Nothing} = nothing: Real number to set the tolerance for any OED-solver

dt::Union{Real, Nothing} = nothing: Real number to set the step size of the OED-solver. Defaults to an automatic choice if the method is adaptive. More Info: DifferentialEquations.jl Documentation

solver = nothing: Any Julia-supported OED-solver (default is Tsit5). More Info: DifferentialEquations.jl Documentation

customFx = nothing: [deperecated] Ability to give a custom state derivative function ẋ=f(x,t)

recordValues::fmi2ValueReferenceFormat = nothing: AbstractArray of variables (strings or variableIdentifiers) to record. Results are returned as DiffEqCallbacks.SavedValues

saveat = nothing: Time points to save values at (interpolated). More Info: DifferentialEquations.jl Documentation

x0::Union{AbstractArray{<:Real}, Nothing} = nothing: Stores the specific value of fmi2ScalarVariable containing the modelVariables with the identical fmi2ValueReference to the input variable vr (vr = vrs[i]). And is therefore passed within prepareSolveFMU to fmi2Set , to set the start state.

setup::Bool = true: Boolean, if FMU should be setup (default: setup=true)

reset::Union{Bool, Nothing} = nothing: Boolean, if FMU should be reset before simulation (default: reset:=auto)

instantiate::Union{Bool, Nothing} = nothing: Boolean value that decides whether to create a new instance of the specified fmu.

freeInstance::Union{Bool, Nothing} = nothing: Boolean value that determines whether to dispose of the given instance, unload the loaded model, and free the allocated memory and other resources allocated by the FMU interface functions.

terminate::Union{Bool, Nothing} = nothing: Boolean value that tells the FMU that the simulation run will be aborted.

inputValueReferences::fmi2ValueReferenceFormat = nothing: AbstractArray of input variables (strings or variableIdentifiers) to set at every simulation step

inputFunction = nothing: Function to retrieve the values to set the inputs to

parameters::Union{Dict{<:Any, <:Any}, Nothing} = nothing: Dictionary of parameter variables (strings or variableIdentifiers) and values (Real, Integer, Boolean, String) to set parameters during initialization

dtmax::Union{Real, Nothing} = nothing: Real number for setting maximum dt for adaptive timestepping for the ODE solver. The default values are package dependent. More Info: DifferentialEquations.jl Documentation

callbacks = []: custom callbacks to add.

showProgress::Bool = true: Boolean value that determines whether a progress bar is generated for a task

kwargs...: Further parameters of already defined functions solve(args..., kwargs...) from the library DifferentialEquations.jl

Returns

If keyword recordValues is not set, a struct of type ODESolution.

If keyword recordValues is set, a tuple of type (ODESolution, DiffEqCallbacks.SavedValues).

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.22]: 2.1.4 Inquire Platform and Version Number of Header Files

FMISpec2.0.2[p.16]: 2.1.2 Platform Dependent Definitions

See also fmi2SimulateME fmi2SimulateCS, fmi2Simulate, fmi2Struct, FMU2, FMU2Component.

FMI.fmiUnload — Function

fmiUnload(fmu::Union{FMU2, FMU3})

Unloads the FMU and all its instances and frees the allocated memory.

Arguments

fmu::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

fmu::FMU3: Mutable struct representing a FMU and all it instantiated instances in the FMI 3.0 Standard.

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.22]: 2.1.4 Inquire Platform and Version Number of Header Files

FMISpec2.0.2[p.16]: 2.1.2 Platform Dependent Definitions

See also fmi2Unload, fmi3Unload.

FMI.fmiReload — Function

fmiReload(fmu::Union{FMU2, FMU3})

Reloads the FMU-binary. This is useful, if the FMU does not support a clean reset implementation.

Arguments

fmu::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

fmu::FMU3: Mutable struct representing an FMU in the FMI 3.0 Standard.

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.22]: 2.1.4 Inquire Platform and Version Number of Header Files

FMISpec2.0.2[p.16]: 2.1.2 Platform Dependent Definitions

See also fmi2Reload, fmi3Reload.

Handling Value References

FMI.fmiStringToValueReference — Function

fmiStringToValueReference(dataStruct::Union{FMU2, fmi2ModelDescription}, identifier::Union{String, AbstractArray{String}})

Wrapper for fmi2StringToValueReference.

See also fmi2ValueReferenceToString.

fmiStringToValueReference(dataStruct::Union{FMU3, fmi3ModelDescription}, identifier::Union{String, AbstractArray{String}})

Wrapper for fmi3StringToValueReference.

See also fmi3ValueReferenceToString.

FMI.fmiGetStartValue — Function

fmiGetStartValue(s::fmi2Struct, vr::fmi2ValueReferenceFormat)

Returns the start/default value for a given value reference.

Arguments

md::fmi2ModelDescription: Struct which provides the static information of ModelVariables.

vrs::fmi2ValueReferenceFormat = md.valueReferences: wildcards for how a user can pass a fmi[X]ValueReference (default = md.valueReferences)

More detailed: fmi2ValueReferenceFormat = Union{Nothing, String, Array{String,1}, fmi2ValueReference, Array{fmi2ValueReference,1}, Int64, Array{Int64,1}, Symbol}

Returns

starts::Array{fmi2ValueReferenceFormat}: start/default value for a given value reference

External/additional functions

FMI.fmiInfo — Function

fmiInfo(str::fmi2Struct)

Prints FMU-specific information into the REPL.

Arguments

str::Union{fmi2Struct, fmi3Struct}: Representative for an FMU in the FMI 2.0.2 Standard or FMI 3.0 Standard. Other notation:

More detailed: fmi2Struct = Union{FMU2, FMU2Component} More detailed: fmi3StructMD = Union{FMU3, FMU3Instance, fmi3ModelDescription}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

str::FMU3: Mutable struct representing an FMU in the FMI 3.0 Standard.

str::FMU3Instance: Mutable struct represents a pointer to an FMU specific data structure that contains the information needed. Also in FMI 3.0 Standard.

Returns

Prints FMU related information.

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec3.0 Link: https://fmi-standard.org/

See also fmi2Info, fmi2Struct, FMU2, FMU2Component, fmi3Info, fmi3Struct, FMU3, FMU3Instance.

FMI.fmiSet — Function

fmiSet(str::fmi2Struct, comp::FMU2Component, vrs::fmi2ValueReferenceFormat, srcArray::AbstractArray; filter=nothing)

Arguments

str::fmi2Struct: Representative for an FMU in the FMI 2.0.2 Standard.

More detailed: fmi2Struct = Union{FMU2, FMU2Component}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

srcArray::AbstractArray: Stores the specific value of fmi2ScalarVariable containing the modelVariables with the identical fmi2ValueReference to the input variable vr (vr = vrs[i]). srcArray has the same length as vrs.

Keywords

filter=nothing: whether the individual values of "fmi2ScalarVariable" are to be stored

Returns

retcodes::Array{fmi2Status}: Returns an array of length length(vrs) with Type fmi2Status. Type fmi2Status is an enumeration and indicates the success of the function call.

More detailed:

fmi2OK: all well

fmi2Warning: things are not quite right, but the computation can continue

fmi2Discard: if the slave computed successfully only a subinterval of the communication step

fmi2Error: the communication step could not be carried out at all

fmi2Fatal: if an error occurred which corrupted the FMU irreparably

fmi2Pending: this status is returned if the slave executes the function asynchronously

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.23]: 2.1.6 Initialization, Termination, and Resetting an FMU

FMISpec2.0.2[p.18]: 2.1.3 Status Returned by Functions

FMI.fmiGet — Function

fmiGet(str::fmi2Struct, comp::FMU2Component, vrs::fmi2ValueReferenceFormat)

Returns the specific value of fmi2ScalarVariable containing the modelVariables with the identical fmi2ValueReference in an array.

Arguments

str::fmi2Struct: Representative for an FMU in the FMI 2.0.2 Standard.

More detailed: fmi2Struct = Union{FMU2, FMU2Component}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

vrs::fmi2ValueReferenceFormat: wildcards for how a user can pass a fmi[X]ValueReference

More detailed: fmi2ValueReferenceFormat = Union{Nothing, String, Array{String,1}, fmi2ValueReference, Array{fmi2ValueReference,1}, Int64, Array{Int64,1}, Symbol}

Returns

dstArray::Array{Any,1}(undef, length(vrs)): Stores the specific value of fmi2ScalarVariable containing the modelVariables with the identical fmi2ValueReference to the input variable vr (vr = vrs[i]). dstArray is a 1-Dimensional Array that has the same length as vrs.

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.23]: 2.1.6 Initialization, Termination, and Resetting an FMU

FMISpec2.0.2[p.18]: 2.1.3 Status Returned by Functions

FMI.fmiGet! — Function

fmiGet!(str::fmi2Struct, comp::FMU2Component, vrs::fmi2ValueReferenceFormat, dstArray::AbstractArray)

Stors the specific value of fmi2ScalarVariable containing the modelVariables with the identical fmi2ValueReference in an array.

Arguments

str::fmi2Struct: Representative for an FMU in the FMI 2.0.2 Standard.

More detailed: fmi2Struct = Union{FMU2, FMU2Component}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

dstArray::AbstractArray: Stores the specific value of fmi2ScalarVariable containing the modelVariables with the identical fmi2ValueReference to the input variable vr (vr = vrs[i]). dstArray has the same length as vrs.

Returns

retcodes::Array{fmi2Status}: Returns an array of length length(vrs) with Type fmi2Status. Type fmi2Status is an enumeration and indicates the success of the function call.

More detailed:

fmi2OK: all well

fmi2Warning: things are not quite right, but the computation can continue

fmi2Discard: if the slave computed successfully only a subinterval of the communication step

fmi2Error: the communication step could not be carried out at all

fmi2Fatal: if an error occurred which corrupted the FMU irreparably

fmi2Pending: this status is returned if the slave executes the function asynchronously

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.23]: 2.1.6 Initialization, Termination, and Resetting an FMU

FMISpec2.0.2[p.18]: 2.1.3 Status Returned by Functions

FMI.fmiGetNumberOfStates — Function

fmiGetNumberOfStates(str::Union{fmi2Struct, fmi3Struct})

Returns the number of states of the FMU.

Arguments

str::Union{fmi2Struct, fmi3Struct}: Representative for an FMU in the FMI 2.0.2 Standard or FMI 3.0 Standard. Other notation:

More detailed: fmi2Struct = Union{FMU2, FMU2Component} More detailed: fmi3StructMD = Union{FMU3, FMU3Instance, fmi3ModelDescription}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

str::FMU3: Mutable struct representing an FMU in the FMI 3.0 Standard.

str::FMU3Instance: Mutable struct represents a pointer to an FMU specific data structure that contains the information needed. Also in FMI 3.0 Standard.

Returns

Returns the length of the md.valueReferences::Array{fmi2ValueReference} corresponding to the number of states of the FMU.

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.22]: 2.1.4 Inquire Platform and Version Number of Header Files

FMISpec2.0.2[p.16]: 2.1.2 Platform Dependent Definitions

See also fmi2GetNumberOfStates, fmi2Struct, FMU2, FMU2Component, , fmi3Struct, FMU3, FMU3Instance.

FMI.fmiCanGetSetState — Function

fmiCanGetSetState(str::Union{fmi2StructMD, fmi3StructMD})

Returns true, if the FMU supports the getting/setting of states

Arguments

str::Union{fmi2StructMD, fmi3StructMD}: Representative for an FMU in the FMI 2.0.2 Standard or FMI 3.0 Standard. Other notation:

More detailed: fmi2StructMD = Union{FMU2, FMU2Component, fmi2ModelDescription} More detailed: fmi3StructMD = Union{FMU3, FMU3Instance, fmi3ModelDescription}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

str::fmi2ModelDescription: Struct witch provides the static information of ModelVariables.

str::FMU3: Mutable struct representing an FMU in the FMI 3.0 Standard.

str::FMU3Instance: Mutable struct represents a pointer to an FMU specific data structure that contains the information needed. Also in FMI 3.0 Standard.

str::fmi3ModelDescription: Struct witch provides the static information of ModelVariables.

Returns

::Bool: The function fmi2CanGetSetState returns True, if the FMU supports the getting/setting of states.

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.22]: 2.1.4 Inquire Platform and Version Number of Header Files

FMISpec2.0.2[p.16]: 2.1.2 Platform Dependent Definitions

See also fmi2CanGetSetState, fmi2StructMD, FMU2, FMU2Component, fmi2ModelDescription, fmi3CanGetSetState, fmi3StructMD, FMU3, FMU3Instance, fmi3ModelDescription.

FMI.fmiGetState — Function

fmiGetState(str::fmi2Struct)

Wrapper for fmi2GetFMUstate.

See also fmiSetState, fmiFreeState!.

fmiGetState(str::fmi3Struct)

Wrapper for fmi3GetFMUState.

See also fmiSetState, fmiFreeState!.

FMI.fmiSetState — Function

fmiSetState(str::fmi2Struct, args...; kwargs...)

Wrapper for fmi2SetFMUstate.

See also fmiGetState, fmiFreeState!.

fmiSetState(str::fmi3Struct, args...; kwargs...)

Wrapper for fmi3SetFMUState.

See also fmiGetState, fmiFreeState!.

FMI.fmiFreeState! — Function

fmiFreeState!(str::fmi2Struct, args...; kwargs...)

Wrapper for fmi2FreeFMUstate!.

See also fmiSetState, fmiGetState.

fmiFreeState!(str::fmi3Struct, args...; kwargs...)

Wrapper for fmi3FreeFMUState!.

See also fmiSetState, fmiGetState.

FMI.fmiGetDependencies — Function

fmiGetDependencies(fmu::FMU2)

Wrapper for fmi2GetDependencies.

fmiGetDependencies(fmu::FMU3)

Wrapper for fmi3GetDependencies.

FMI.fmiProvidesDirectionalDerivative — Function

fmiProvidesDirectionalDerivative(str::Union{fmi2StructMD, fmi3StructMD})

Returns true, if the FMU provides directional derivatives

Arguments

str::Union{fmi2StructMD, fmi3StructMD}: Representative for an FMU in the FMI 2.0.2 Standard or FMI 3.0 Standard. Other notation:

More detailed: fmi2StructMD = Union{FMU2, FMU2Component, fmi2ModelDescription} More detailed: fmi3StructMD = Union{FMU3, FMU3Instance, fmi3ModelDescription}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

str::fmi2ModelDescription: Struct witch provides the static information of ModelVariables.

str::FMU3: Mutable struct representing an FMU in the FMI 3.0 Standard.

str::FMU3Instance: Mutable struct represents a pointer to an FMU specific data structure that contains the information needed. Also in FMI 3.0 Standard.

str::fmi3ModelDescription: Struct witch provides the static information of ModelVariables.

Returns

::Bool: The function fmi2ProvidesDirectionalDerivative returns True, if the FMU provides directional derivatives.

See also fmi2ProvidesDirectionalDerivative, fmi2StructMD, FMU2, FMU2Component, fmi2ModelDescription, fmi3ProvidesDirectionalDerivatives, fmi3StructMD, FMU3, FMU3Instance, fmi3ModelDescription.

FMI.fmiGetModelName — Function

fmiGetModelName(str::fmi2StructMD)

Return the modelName from str's model description.

Wrapper for fmi2GetModelName.

fmiGetModelName(str::fmi3StructMD)

Return the modelName from str's model description.

Wrapper for fmi3GetModelName.

FMI.fmiGetGUID — Function

fmiGetGUID(str::fmi2StructMD)

Return the guid from str's model description.

Wrapper for fmi2GetGUID.

FMI.fmiIsCoSimulation — Function

fmiIsCoSimulation(str::Union{fmi2StructMD, fmi3StructMD})

Returns true, if the FMU supports co simulation

Arguments

str::Union{fmi2StructMD, fmi3StructMD}: Representative for an FMU in the FMI 2.0.2 Standard or FMI 3.0 Standard. Other notation:

More detailed: fmi2StructMD = Union{FMU2, FMU2Component, fmi2ModelDescription} More detailed: fmi3StructMD = Union{FMU3, FMU3Instance, fmi3ModelDescription}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

str::fmi2ModelDescription: Struct witch provides the static information of ModelVariables.

str::FMU3: Mutable struct representing an FMU in the FMI 3.0 Standard.

str::FMU3Instance: Mutable struct represents a pointer to an FMU specific data structure that contains the information needed. Also in FMI 3.0 Standard.

str::fmi3ModelDescription: Struct witch provides the static information of ModelVariables.

Returns

::Bool: The function fmi2IsCoSimulation returns True, if the FMU supports co simulation

See also fmi2IsCoSimulation, fmi2StructMD, FMU2, FMU2Component, fmi2ModelDescription, fmi3IsCoSimulation, fmi3StructMD, FMU3, FMU3Instance, fmi3ModelDescription.

FMI.fmiIsModelExchange — Function

fmiIsModelExchange(str::Union{fmi2StructMD, fmi3StructMD})

Returns true, if the FMU supports model exchange

Arguments

str::Union{fmi2StructMD, fmi3StructMD}: Representative for an FMU in the FMI 2.0.2 Standard or FMI 3.0 Standard. Other notation:

More detailed: fmi2StructMD = Union{FMU2, FMU2Component, fmi2ModelDescription} More detailed: fmi3StructMD = Union{FMU3, FMU3Instance, fmi3ModelDescription}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

str::fmi2ModelDescription: Struct witch provides the static information of ModelVariables.

str::FMU3: Mutable struct representing an FMU in the FMI 3.0 Standard.

str::FMU3Instance: Mutable struct represents a pointer to an FMU specific data structure that contains the information needed. Also in FMI 3.0 Standard.

str::fmi3ModelDescription: Struct witch provides the static information of ModelVariables.

Returns

::Bool: The function fmi2IsModelExchange returns True, if the FMU supports model exchange.

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.22]: 2.1.4 Inquire Platform and Version Number of Header Files

FMISpec2.0.2[p.16]: 2.1.2 Platform Dependent Definitions

See also fmi2IsModelExchange, fmi2StructMD, FMU2, FMU2Component, fmi2ModelDescription, fmi3IsModelExchange, fmi3StructMD, FMU3, FMU3Instance, fmi3ModelDescription.

FMI.fmiIsScheduledExecution — Function

fmiIsScheduledExecution(str::fmi3StructMD)

Returns true, if the FMU supports scheduled execution

Arguments

str::fmi3StructMD: Representative for an FMU in the FMI 3.0 Standard. Other notation:

More detailed: fmi3StructMD = Union{FMU3, FMU3Instance, fmi3ModelDescription}

str::FMU3: Mutable struct representing an FMU in the FMI 3.0 Standard.

str::FMU3Instance: Mutable struct represents a pointer to an FMU specific data structure that contains the information needed. Also in FMI 3.0 Standard.

str::fmi3ModelDescription: Struct witch provides the static information of ModelVariables.

Returns

::Bool: The function fmi3IsScheduledExecution returns True, if the FMU supports scheduled execution.

See also fmi3IsScheduledExecution, fmi3StructMD, FMU3, FMU3Instance, fmi3ModelDescription.

Visualize simulation results

FMI.fmiPlot — Function

fmiPlot(solution::FMUSolution; kwargs...)

Create a figure and fmiPlot!'s the solution of a FMU simulation with the kwargs into it and returns the figure. (requires Package Plots in Julia Environment)

See also fmiPlot!

FMI.fmiPlot! — Function

fmiPlot!(fig::Plots.Plot, solution::FMUSolution; 
            [states::Union{Bool, Nothing}=nothing,
            values::Union{Bool, Nothing}=nothing,
            stateEvents::Union{Bool, Nothing}=nothing,
            timeEvents::Union{Bool, Nothing}=nothing,
            stateIndices=nothing,
            valueIndices=nothing,
            maxLabelLength=64,
            plotkwargs...])

Plot the solution of a FMU simulation into fig and return the figure.

Arguments

fig::Plots.Plot: Figure to plot into

solution::FMUSolution: Struct containing information about the solutions values, success, states and events of a specific FMU simulation.

states::Union{Bool, Nothing}=nothing: controls if states should be plotted (default = nothing: plot states from solution, as long as they exist)

values::Union{Bool, Nothing}=nothing: controls if values should be plotted (default = nothing: plot values from solution, as long as they exist)

stateEvents::Union{Bool, Nothing}=nothing: controls if stateEvents should be plotted (default = nothing: plot stateEvents from solution, if at least one and at most 100 exist)

timeEvents::Union{Bool, Nothing}=nothing: controls if timeEvents should be plotted (default = nothing: plot timeEvents from solution, if at least one and at most 100 exist)

stateIndices=nothing: controls which states will be plotted by index in state vector (default = nothing: plot all states)

valueIndices=nothing: controls which values will be plotted by index (default = nothing: plot all values)

maxLabelLength=64: controls the maximum length for legend labels (too long labels are cut from front)

plotkwargs...: Arguments, that are passed on to Plots.plot!

See also fmiPlot

RecipesBase.plot — Function

Plots.plot(solution::FMUSolution; kwargs...)
Plots.plot!(fig::Plots.Plot, solution::FMUSolution; kwargs...)

Plot FMUs using the original plot-command from Plots.

See also fmiPlot, fmiPlot!.

Save/load simulation results

FMI.fmiSaveSolution — Function

fmiSaveSolution(solution::FMUSolution, filepath::AbstractString [; keyword="solution"])

Save a solution of an FMU simulation at filepath.

Currently .mat, .jld2 and .csv are supported for saving and selected by the ending of filepath. For JLD2 the keyword is used as key. Loading a FMUSolution into FMI.jl is currently only possible for .jld2 files.

See also fmiSaveSolutionCSV, fmiSaveSolutionMAT, fmiSaveSolutionJLD2, fmiLoadSolutionJLD2.

FMI.fmiSaveSolutionJLD2 — Function

fmiSaveSolutionJLD2(solution::FMUSolution, filepath::AbstractString; keyword="solution")

Save a solution of an FMU simulation under keyword in a jld2 file at filepath. (requires Package JLD2 in Julia Environment)

See also fmiSaveSolutionCSV, fmiSaveSolutionMAT, fmiLoadSolutionJLD2.

FMI.fmiSaveSolutionMAT — Function

fmiSaveSolutionMAT(solution::FMUSolution, filepath::AbstractString)

Save a solution of an FMU simulation as mat file at filepath. (requires Package MAT in Julia Environment)

See also fmiSaveSolutionCSV, fmiSaveSolutionJLD2, fmiLoadSolutionJLD2.

FMI.fmiSaveSolutionCSV — Function

fmiSaveSolutionCSV(solution::FMUSolution, filepath::AbstractString)

Save a solution of an FMU simulation as csv file at filepath. (requires Package CSV in Julia Environment)

See also fmiSaveSolutionMAT, fmiSaveSolutionJLD2, fmiLoadSolutionJLD2.

FMI.fmiLoadSolution — Function

fmiLoadSolution(filepath::AbstractString; keyword="solution")

Wrapper for fmiLoadSolutionJLD2.

FMI.fmiLoadSolutionJLD2 — Function

fmiLoadSolutionJLD2(filepath::AbstractString; keyword="solution")

Load a FMUSolution from jld2 file at filepath using keyword as jld2 keyword. (requires Package JLD2 in Julia Environment)

See also fmiSaveSolutionCSV, fmiSaveSolutionMAT, fmiSaveSolutionJLD2.

FMI2 specific

FMI.fmi2Info — Function

 fmi2Info(fmu::FMU2)

Print information about the fmu.

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.22]: 2.1.4 Inquire Platform and Version Number of Header Files

FMISpec2.0.2[p.16]: 2.1.2 Platform Dependent Definitions

FMI.fmi2Simulate — Function

fmi2Simulate(fmu::FMU2, args...; kwargs...)

Wrapper for fmi2Simulate(fmu::FMU2, c::Union{FMU2Component, Nothing}, args...; kwargs...) without a provided FMU2Component. (Component c gets passed as nothing)

fmi2Simulate(c::FMU2Component, tspan::Union{Tuple{Float64, Float64}, Nothing}=nothing; kwargs...)

Wrapper for fmi2Simulate(fmu::FMU2, c::Union{FMU2Component, Nothing}, tspan::Union{Tuple{Float64, Float64}, Nothing}=nothing; kwargs...) without a provided FMU2. (FMU2 fmu is taken from c)

fmi2Simulate(args...)

Starts a simulation of the FMU2 for the matching type (fmi2SimulateCS(args...) or fmi2SimulateME(args...)); if both types are available, CS is preferred.

See also fmi2SimulateCS, fmi2SimulateME.

FMI.fmi2SimulateME — Function

fmi2SimulateME(fmu::FMU2, args...; kwargs...)

Wrapper for fmi2SimulateME(fmu::FMU2, c::Union{FMU2Component, Nothing}, args...; kwargs...) without a provided FMU2Component. (Component c gets passed as nothing)

fmi2SimulateME(c::FMU2Component, tspan::Union{Tuple{Float64, Float64}, Nothing}=nothing; kwargs...)

Wrapper for fmi2SimulateME(fmu::FMU2, c::Union{FMU2Component, Nothing}, tspan::Union{Tuple{Float64, Float64}, Nothing}=nothing; kwargs...) without a provided FMU2. (FMU2 fmu is taken from c)

fmi2SimulateME(fmu::FMU2, 
            c::Union{FMU2Component, Nothing}=nothing, 
            tspan::Union{Tuple{Float64, Float64}, Nothing}=nothing;
            [tolerance::Union{Real, Nothing} = nothing,
            dt::Union{Real, Nothing} = nothing,
            solver = nothing,
            customFx = nothing,
            recordValues::fmi2ValueReferenceFormat = nothing,
            recordEventIndicators::Union{AbstractArray{<:Integer, 1}, UnitRange{<:Integer}, Nothing} = nothing,
            recordEigenvalues::Bool=false,
            saveat = nothing,
            x0::Union{AbstractArray{<:Real}, Nothing} = nothing,
            setup::Union{Bool, Nothing} = nothing,
            reset::Union{Bool, Nothing} = nothing,
            instantiate::Union{Bool, Nothing} = nothing,
            freeInstance::Union{Bool, Nothing} = nothing,
            terminate::Union{Bool, Nothing} = nothing,
            inputValueReferences::fmi2ValueReferenceFormat = nothing,
            inputFunction = nothing,
            parameters::Union{Dict{<:Any, <:Any}, Nothing} = nothing,
            dtmax::Union{Real, Nothing} = nothing,
            callbacksBefore = [],
            callbacksAfter = [],
            showProgress::Bool = true,
            kwargs...])

Simulate ME-FMU for the given simulation time interval.

State- and Time-Events are handled correctly.

Arguments

fmu::FMU2: Mutable struct representing a FMU and all it instantiated instances.

c::Union{FMU2Component, Nothing}=nothing: Mutable struct representing an instantiated instance of a FMU.

tspan::Union{Tuple{Float64, Float64}, Nothing}=nothing: Simulation-time-span as tuple (default = nothing: use default value from fmu's model description or (0.0, 1.0))

tolerance::Union{Real, Nothing} = nothing: tolerance for the solver (default = nothing: use default value from fmu's model description or -if not available- default from DifferentialEquations.jl)

dt::Union{Real, Nothing} = nothing: stepszie for the solver (default = nothing: use default value from fmu's model description or -if not available- default from DifferentialEquations.jl)

solver = nothing: Any Julia-supported ODE-solver (default = nothing: use DifferentialEquations.jl default solver)

customFx: [deprecated] custom state derivative function ẋ=f(x,t)

recordValues::fmi2ValueReferenceFormat = nothing: Array of variables (Strings or variableIdentifiers) to record. Results are returned as DiffEqCallbacks.SavedValues

recordEventIndicators::Union{AbstractArray{<:Integer, 1}, UnitRange{<:Integer}, Nothing} = nothing: Array or Range of event indicators to record

recordEigenvalues::Bool=false: compute and record eigenvalues

saveat = nothing: Time points to save (interpolated) values at (default = nothing: save at each solver timestep)

x0::Union{AbstractArray{<:Real}, Nothing} = nothing: inital fmu State (default = nothing: use current or default-inital fmu state)

setup::Union{Bool, Nothing} = nothing: call fmi2SetupExperiment, fmi2EnterInitializationMode and fmi2ExitInitializationMode before each step (default = nothing: use value from fmu's FMU2ExecutionConfiguration)

reset::Union{Bool, Nothing} = nothing: call fmi2Reset before each step (default = nothing: use value from fmu's FMU2ExecutionConfiguration)

instantiate::Union{Bool, Nothing} = nothing: call fmi2Instantiate! before each step (default = nothing: use value from fmu's FMU2ExecutionConfiguration)

freeInstance::Union{Bool, Nothing} = nothing: call fmi2FreeInstance after each step (default = nothing: use value from fmu's FMU2ExecutionConfiguration)

terminate::Union{Bool, Nothing} = nothing: call fmi2Terminate after each step (default = nothing: use value from fmu's FMU2ExecutionConfiguration)

inputValueReferences::fmi2ValueReferenceFormat = nothing: Input variables (Strings or variableIdentifiers) to set at each simulation step

inputFunction = nothing: Function to get values for the input variables at each simulation step.

Pattern [c: current component, u: current state ,t: current time, returning array of values to be passed to fmi2SetReal(..., inputValueReferences, inputFunction(...))]:

inputFunction(t::fmi2Real)

inputFunction(c::FMU2Component, t::fmi2Real)

inputFunction(c::FMU2Component, u::Union{AbstractArray{fmi2Real,1}, Nothing})

inputFunction(u::Union{AbstractArray{fmi2Real,1}, Nothing}, t::fmi2Real)

inputFunction(c::FMU2Component, u::Union{AbstractArray{fmi2Real,1}, Nothing}, t::fmi2Real)

parameters::Union{Dict{<:Any, <:Any}, Nothing} = nothing: Dict of parameter variables (strings or variableIdentifiers) and values (Real, Integer, Boolean, String) to set parameters during initialization

dtmax::Union{Real, Nothing} = nothing: sets the maximum stepszie for the solver (default = nothing: use (Simulation-time-span-length)/100.0)

callbacksBefore = [], callbacksAfter = []: functions that are to be called before and after internal time-event-, state-event- and step-event-callbacks are called

showProgress::Bool = true: print simulation progressmeter in REPL

kwargs...: keyword arguments that get passed onto the solvers solve call

Returns:

If keyword recordValues has value nothing, a struct of type ODESolution.

If keyword recordValues is set, a tuple of type (ODESolution, DiffEqCallbacks.SavedValues).

See also fmi2Simulate, fmi2SimulateCS.

FMI.fmi2SimulateCS — Function

fmi2SimulateCS(fmu::FMU2, args...; kwargs...)

Wrapper for fmi2SimulateCS(fmu::FMU2, c::Union{FMU2Component, Nothing}, args...; kwargs...) without a provided FMU2Component. (Component c gets passed as nothing)

fmi2SimulateCS(c::FMU2Component, tspan::Union{Tuple{Float64, Float64}, Nothing}=nothing; kwargs...)

Wrapper for fmi2SimulateCS(fmu::FMU2, c::Union{FMU2Component, Nothing}, tspan::Union{Tuple{Float64, Float64}, Nothing}=nothing; kwargs...) without a provided FMU2. (FMU2 fmu is taken from c)

fmi2SimulateCS(fmu::FMU2, 
            c::Union{FMU2Component, Nothing}=nothing, 
            tspan::Union{Tuple{Float64, Float64}, Nothing}=nothing;
            [tolerance::Union{Real, Nothing} = nothing,
            dt::Union{Real, Nothing} = nothing,
            recordValues::fmi2ValueReferenceFormat = nothing,
            saveat = [],
            setup::Union{Bool, Nothing} = nothing,
            reset::Union{Bool, Nothing} = nothing,
            instantiate::Union{Bool, Nothing} = nothing,
            freeInstance::Union{Bool, Nothing} = nothing,
            terminate::Union{Bool, Nothing} = nothing,
            inputValueReferences::fmi2ValueReferenceFormat = nothing,
            inputFunction = nothing,
            showProgress::Bool=true,
            parameters::Union{Dict{<:Any, <:Any}, Nothing} = nothing])

Simulate CS-FMU for the given simulation time interval.

Arguments

fmu::FMU2: Mutable struct representing a FMU and all it instantiated instances.

c::Union{FMU2Component, Nothing}=nothing: Mutable struct representing an instantiated instance of a FMU.

tspan::Union{Tuple{Float64, Float64}, Nothing}=nothing: Simulation-time-span as tuple (default = nothing: use default value from fmu's model description or (0.0, 1.0))

tolerance::Union{Real, Nothing} = nothing: tolerance for the solver (default = nothing: use default value from fmu's model description or 0.0)

dt::Union{Real, Nothing} = nothing: stepszie for the solver (default = nothing: use default value from fmu's model description or 1e-3)

recordValues::fmi2ValueReferenceFormat = nothing: Array of variables (Strings or variableIdentifiers) to record. Results are returned as DiffEqCallbacks.SavedValues

saveat = nothing: Time points to save values at (default = nothing: save at each communication timestep)

setup::Union{Bool, Nothing} = nothing: call fmi2SetupExperiment, fmi2EnterInitializationMode and fmi2ExitInitializationMode before each step (default = nothing: use value from fmu's FMU2ExecutionConfiguration)

reset::Union{Bool, Nothing} = nothing: call fmi2Reset before each step (default = nothing: use value from fmu's FMU2ExecutionConfiguration)

instantiate::Union{Bool, Nothing} = nothing: call fmi2Reset before each step (default = nothing: use value from fmu's FMU2ExecutionConfiguration)

freeInstance::Union{Bool, Nothing} = nothing: call fmi2FreeInstance after each step (default = nothing: use value from fmu's FMU2ExecutionConfiguration)

terminate::Union{Bool, Nothing} = nothing: call fmi2Terminate after each step (default = nothing: use value from fmu's FMU2ExecutionConfiguration)

inputValueReferences::fmi2ValueReferenceFormat = nothing: Input variables (Strings or variableIdentifiers) to set at each communication step

inputFunction = nothing: Function to get values for the input variables at each communication step.

Pattern [c: current component, t: current time, returning array of values to be passed to fmi2SetReal(..., inputValueReferences, inputFunction(...))]:

inputFunction(t::fmi2Real)

inputFunction(c::FMU2Component, t::fmi2Real)

showProgress::Bool = true: print simulation progressmeter in REPL

parameters::Union{Dict{<:Any, <:Any}, Nothing} = nothing: Dict of parameter variables (strings or variableIdentifiers) and values (Real, Integer, Boolean, String) to set parameters during initialization

Returns:

fmusol::FMU2Solution, containing bool fmusol.success and if keyword recordValues is set, the saved values as fmusol.values.

See also fmi2Simulate, fmi2SimulateME.

FMI.fmi2VariableDependsOnVariable — Function

fmi2VariableDependsOnVariable(fmu::FMU2, vr1::fmi2ValueReference, vr2::fmi2ValueReference)

Return the dependence of the variable described by vr1 on another variable described by vr2 based on the model description of the fmu.

See also fmi2GetDependencies.

FMI.fmi2GetDependencies — Function

fmi2GetDependencies(fmu::FMU2)

Build dependency Matrix{Union{fmi2DependencyKind, Nothing}} of dimension n x n for fast look-ups on dependencies between value references (n is number of states of the fmu).

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.22]: 2.1.4 Inquire Platform and Version Number of Header Files

FMISpec2.0.2[p.16]: 2.1.2 Platform Dependent Definitions

See also fmi2PrintDependencies, fmi2VariableDependsOnVariable.

FMI.fmi2PrintDependencies — Function

fmi2PrintDependencies(fmu::FMU2)

Print the dependency matrix for fmu as returned by fmi2GetDependencies.

See also fmi2GetDependencies.

FMI3 specific

FMI.fmi3Info — Function

 fmi3Info(fmu::FMU3)

Print information about the fmu.

FMI.fmi3Simulate — Function

fmi3Simulate(fmu::FMU3, args...; kwargs...)

Wrapper for fmi3Simulate(fmu::FMU3, c::Union{FMU3Instance, Nothing}, args...; kwargs...) without a provided FMU3Instance. (Instance c gets passed as nothing)

fmi3Simulate(c::FMU3Instance, t_start::Union{Real, Nothing} = nothing, t_stop::Union{Real, Nothing} = nothing; kwargs...)

Wrapper for fmi3Simulate(fmu::FMU3, c::Union{FMU3Instance, Nothing}=nothing, t_start::Union{Real, Nothing} = nothing, t_stop::Union{Real, Nothing} = nothing; kwargs...) without a provided FMU3. (FMU3 fmu is taken from c)

fmi3Simulate(args...)

Starts a simulation of the FMU3 for the matching type (fmi3SimulateCS(args...), fmi3SimulateME(args...) or fmi3SimulateSE(args...)); if multiple types are available, CS is preferred over ME, over SE.

See also fmi3SimulateCS, fmi3SimulateME, fmi3SimulateSE.

FMI.fmi3SimulateME — Function

fmi3SimulateME(fmu::FMU3, args...; kwargs...)

Wrapper for fmi3SimulateME(fmu::FMU3, c::Union{FMU3Instance, Nothing}, args...; kwargs...) without a provided FMU3Instance. (Instance c gets passed as nothing)

fmi3SimulateME(c::FMU3Instance, t_start::Union{Real, Nothing} = nothing, t_stop::Union{Real, Nothing} = nothing; kwargs...)

Wrapper for fmi3SimulateME(fmu::FMU3, c::Union{FMU3Instance, Nothing}, t_start::Union{Real, Nothing} = nothing, t_stop::Union{Real, Nothing} = nothing; kwargs...) without a provided FMU3. (FMU3 fmu is taken from c)

fmi3SimulateME(fmu::FMU3, 
            c::Union{FMU3Instance, Nothing}=nothing, 
            t_start::Union{Real, Nothing} = nothing, 
            t_stop::Union{Real, Nothing} = nothing;
            [tolerance::Union{Real, Nothing} = nothing,
            dt::Union{Real, Nothing} = nothing,
            solver = nothing,
            customFx = nothing,
            recordValues::fmi3ValueReferenceFormat = nothing,
            saveat = nothing,
            x0::Union{AbstractArray{<:Real}, Nothing} = nothing,
            setup::Union{Bool, Nothing} = nothing,
            reset::Union{Bool, Nothing} = nothing,
            instantiate::Union{Bool, Nothing} = nothing,
            freeInstance::Union{Bool, Nothing} = nothing,
            terminate::Union{Bool, Nothing} = nothing,
            inputValueReferences::fmi3ValueReferenceFormat = nothing,
            inputFunction = nothing,
            parameters::Union{Dict{<:Any, <:Any}, Nothing} = nothing,
            dtmax::Union{Real, Nothing} = nothing,
            callbacks = [],
            showProgress::Bool = true,
            kwargs...])

Simulate ME-FMU for the given simulation time interval.

State- and Time-Events are handled correctly.

Arguments

fmu::FMU3: Mutable struct representing a FMU and all it instantiated instances.

c::Union{FMU3Instance, Nothing}=nothing: Mutable struct representing an instantiated instance of a FMU.

t_start::Union{Real, Nothing} = nothing: Simulation-time-span start time (default = nothing: use default value from fmu's model description or 0.0)

t_stop::Union{Real, Nothing} = nothing: Simulation-time-span stop time (default = nothing: use default value from fmu's model description or 1.0)

tolerance::Union{Real, Nothing} = nothing: tolerance for the solver (default = nothing: use default value from fmu's model description or -if not available- default from DifferentialEquations.jl)

dt::Union{Real, Nothing} = nothing: stepszie for the solver (default = nothing: use default value from fmu's model description or -if not available- default from DifferentialEquations.jl)

solver = nothing: Any Julia-supported ODE-solver (default = nothing: use DifferentialEquations.jl default solver)

customFx: [deprecated] custom state derivative function ẋ=f(x,t)

recordValues::fmi3ValueReferenceFormat = nothing: Array of variables (Strings or variableIdentifiers) to record. Results are returned as DiffEqCallbacks.SavedValues

saveat = nothing: Time points to save (interpolated) values at (default = nothing: save at each solver timestep)

x0::Union{AbstractArray{<:Real}, Nothing} = nothing: inital fmu State (default = nothing: use current or default-inital fmu state)

setup::Union{Bool, Nothing} = nothing: call fmi3EnterInitializationMode and fmi3ExitInitializationMode before each step (default = nothing: use value from fmu's FMU3ExecutionConfiguration)

reset::Union{Bool, Nothing} = nothing: call fmi3Reset before each step (default = nothing: use value from fmu's FMU3ExecutionConfiguration)

instantiate::Union{Bool, Nothing} = nothing: call fmi3Instantiate<FMU-Type>! before each step (default = nothing: use value from fmu's FMU3ExecutionConfiguration)

freeInstance::Union{Bool, Nothing} = nothing: call fmi3FreeInstance after each step (default = nothing: use value from fmu's FMU3ExecutionConfiguration)

terminate::Union{Bool, Nothing} = nothing: call fmi3Terminate after each step (default = nothing: use value from fmu's FMU3ExecutionConfiguration)

inputValueReferences::fmi3ValueReferenceFormat = nothing: Input variables (Strings or variableIdentifiers) to set at each simulation step

inputFunction = nothing: Function to get values for the input variables at each simulation step.

Pattern [c: current instance, u: current state ,t: current time, returning array of values to be passed to fmi3SetFloat64(..., inputValueReferences, inputFunction(...))]:

inputFunction(t::fmi3Float64)

inputFunction(c::FMU3Instance, t::fmi3Float64)

inputFunction(c::FMU3Instance, u::Union{AbstractArray{fmi3Float64,1}, Nothing})

inputFunction(u::Union{AbstractArray{fmi3Float64,1}, Nothing}, t::fmi3Float64)

inputFunction(c::FMU3Instance, u::Union{AbstractArray{fmi3Float64,1}, Nothing}, t::fmi3Float64)

parameters::Union{Dict{<:Any, <:Any}, Nothing} = nothing: Dict of parameter variables (strings or variableIdentifiers) and values (Real, Integer, Boolean, String) to set parameters during initialization

dtmax::Union{Real, Nothing} = nothing: sets the maximum stepszie for the solver (default = nothing: use (Simulation-time-span-length)/100.0)

callbacks = []: functions that are to be called at each solver time step

showProgress::Bool = true: print simulation progressmeter in REPL

kwargs...: keyword arguments that get passed onto the solvers solve call

Returns:

If keyword recordValues has value nothing, a struct of type ODESolution.

If keyword recordValues is set, a tuple of type (ODESolution, DiffEqCallbacks.SavedValues).

See also fmi3Simulate, fmi3SimulateCS, fmi3SimulateSE.

FMI.fmi3SimulateSE — Function

fmi3SimulateSE(fmu::FMU3, 
            c::Union{FMU3Instance, Nothing}=nothing, 
            t_start::Union{Real, Nothing} = nothing, 
            t_stop::Union{Real, Nothing} = nothing;
            [tolerance::Union{Real, Nothing} = nothing,
            dt::Union{Real, Nothing} = nothing,
            recordValues::fmi3ValueReferenceFormat = nothing,
            saveat = [],
            setup::Union{Bool, Nothing} = nothing,
            reset::Union{Bool, Nothing} = nothing,
            instantiate::Union{Bool, Nothing} = nothing,
            freeInstance::Union{Bool, Nothing} = nothing,
            terminate::Union{Bool, Nothing} = nothing,
            inputValueReferences::fmi3ValueReferenceFormat = nothing,
            inputFunction = nothing,
            showProgress::Bool=true,
            parameters::Union{Dict{<:Any, <:Any}, Nothing} = nothing])

Simulate SE-FMU; not yet implemented in library

See also fmi3Simulate, fmi3SimulateME, fmi3SimulateCS.

FMI.fmi3SimulateCS — Function

fmi3SimulateCS(fmu::FMU3, args...; kwargs...)

Wrapper for fmi3SimulateCS(fmu::FMU3, c::Union{FMU3Instance, Nothing}, args...; kwargs...) without a provided FMU3Instance. (Instance c gets passed as nothing)

fmi3SimulateCS(c::FMU3Instance, t_start::Union{Real, Nothing} = nothing, t_stop::Union{Real, Nothing} = nothing; kwargs...)

Wrapper for fmi3SimulateCS(fmu::FMU3, c::Union{FMU3Instance, Nothing}=nothing, t_start::Union{Real, Nothing} = nothing, t_stop::Union{Real, Nothing} = nothing; kwargs...) without a provided FMU3. (FMU3 fmu is taken from c)

fmi3SimulateCS(fmu::FMU3, 
            c::Union{FMU3Instance, Nothing}=nothing, 
            t_start::Union{Real, Nothing} = nothing, 
            t_stop::Union{Real, Nothing} = nothing;
            [tolerance::Union{Real, Nothing} = nothing,
            dt::Union{Real, Nothing} = nothing,
            recordValues::fmi3ValueReferenceFormat = nothing,
            saveat = [],
            setup::Union{Bool, Nothing} = nothing,
            reset::Union{Bool, Nothing} = nothing,
            instantiate::Union{Bool, Nothing} = nothing,
            freeInstance::Union{Bool, Nothing} = nothing,
            terminate::Union{Bool, Nothing} = nothing,
            inputValueReferences::fmi3ValueReferenceFormat = nothing,
            inputFunction = nothing,
            showProgress::Bool=true,
            parameters::Union{Dict{<:Any, <:Any}, Nothing} = nothing])

Simulate CS-FMU for the given simulation time interval.

Arguments

fmu::FMU3: Mutable struct representing a FMU and all it instantiated instances.

c::Union{FMU3Instance, Nothing}=nothing: Mutable struct representing an instantiated instance of a FMU.

t_start::Union{Real, Nothing} = nothing: Simulation-time-span start time (default = nothing: use default value from fmu's model description or 0.0)

t_stop::Union{Real, Nothing} = nothing: Simulation-time-span stop time (default = nothing: use default value from fmu's model description or 1.0)

tolerance::Union{Real, Nothing} = nothing: tolerance for the solver (default = nothing: use default value from fmu's model description or 0.0)

dt::Union{Real, Nothing} = nothing: stepszie for the solver (default = nothing: use default value from fmu's model description or 1e-3)

solver = nothing: Any Julia-supported ODE-solver (default = nothing: use DifferentialEquations.jl default solver)

recordValues::fmi3ValueReferenceFormat = nothing: Array of variables (Strings or variableIdentifiers) to record. Results are returned as DiffEqCallbacks.SavedValues

saveat = nothing: Time points to save values at (default = nothing: save at each communication timestep)

setup::Union{Bool, Nothing} = nothing: call fmi3EnterInitializationMode and fmi3ExitInitializationMode before each step (default = nothing: use value from fmu's FMU3ExecutionConfiguration)

reset::Union{Bool, Nothing} = nothing: call fmi3Reset before each step (default = nothing: use value from fmu's FMU3ExecutionConfiguration)

instantiate::Union{Bool, Nothing} = nothing: call fmi3Instantiate<FMU-Type>! before each step (default = nothing: use value from fmu's FMU3ExecutionConfiguration)

freeInstance::Union{Bool, Nothing} = nothing: call fmi3FreeInstance after each step (default = nothing: use value from fmu's FMU3ExecutionConfiguration)

terminate::Union{Bool, Nothing} = nothing: call fmi3Terminate after each step (default = nothing: use value from fmu's FMU3ExecutionConfiguration)

inputValueReferences::fmi3ValueReferenceFormat = nothing: Input variables (Strings or variableIdentifiers) to set at each communication step

inputFunction = nothing: Function to get values for the input variables at each communication step.

Pattern [c: current instance, t: current time, returning array of values to be passed to fmi3SetFloat64(..., inputValueReferences, inputFunction(...))]:

inputFunction(t::fmi3Float64)

inputFunction(c::FMU3Instance, t::fmi3Float64)

showProgress::Bool = true: print simulation progressmeter in REPL

parameters::Union{Dict{<:Any, <:Any}, Nothing} = nothing: Dict of parameter variables (strings or variableIdentifiers) and values (Boolean, String, Float64, ...) to set parameters during initialization

Returns:

fmusol::FMU3Solution, containing bool fmusol.success and if keyword recordValues is set, the saved values as fmusol.values.

See also fmi3Simulate, fmi3SimulateME, fmi3SimulateSE.

FMI.fmi3VariableDependsOnVariable — Function

fmi3VariableDependsOnVariable(fmu::FMU3, vr1::fmi3ValueReference, vr2::fmi3ValueReference)

Return the dependence of the variable described by vr1 on another variable described by vr2 based on the model description of the fmu.

See also fmi3GetDependencies.

FMI.fmi3GetDependencies — Function

fmi3GetDependencies(fmu::FMU3)

Build dependency Matrix{Union{fmi3DependencyKind, Nothing}} of dimension n x n for fast look-ups on dependencies between value references (n is number of states of the fmu).

See also fmi3PrintDependencies, fmi3VariableDependsOnVariable.

FMI.fmi3PrintDependencies — Function

fmi3PrintDependencies(fmu::FMU3)

Print the dependency matrix for fmu as returned by fmi3GetDependencies.

See also fmi3GetDependencies.