Library Functions

Parsing variable names to ValueReferences

FMI.fmiStringToValueReference — Function

fmiStringToValueReference(dataStruct::Union{FMU2, fmi2ModelDescription, FMU3, fmmi3ModelDescription}, identifier::Union{String, AbstractArray{String}})

Returns the ValueReference coresponding to the variable identifier.

Arguments

dataStruct::Union{FMU2, fmi2ModelDescription, FMU3, fmmi3ModelDescription}: Model of the type FMU2/FMU3 or the Model Description of fmi2/fmi3. Same for Model of type FMU3 or the Model Description of fmi3

identifier::Union{String, AbstractArray{String}}: Variable identifier in type String or as a 1-dimensional AbstractArray containing elements of type String

See also fmi2StringToValueReference, fmi3StringToValueReference.

Opening and closing FMUs

FMI.fmiLoad — Function

fmiLoad(pathToFMU::String; unpackPath=nothing, type=nothing)

Load FMUs independent of the FMI version, currently supporting version 2.0.X.

Arguments

pathToFMU::String: String that contains the paths of ziped and unziped FMU folders.

Keywords

unpackPath=nothing: Via optional argument unpackPath, a path to unpack the FMU can be specified (default: system temporary directory).

type::Union{CS, ME} = nothing: Via type, a FMU type can be selected. If none of the unified type set is used, the default value type = nothing will be used.

Returns

Returns the instance of the FMU struct.

See also fmi2Load.

FMI.fmiUnload — Function

fmiUnload(fmu::FMU2)

Unloads the FMU and all its instances and frees the allocated memory.

Arguments

fmu::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

See also fmi2Unload.

Reading the model description

FMI.fmiGetModelName — Function

fmiGetModelName(str::Union{fmi2StructMD, fmi3StructMD})

Returns the tag 'modelName' from the model description.

Arguments

str::Union{fmi2StructMD, fmi3StructMD}: Representative for an FMU in the FMI 2.0.2 Standard or FMI 3.0 Standard. Other notation:

Union{fmi2StructMD, fmi3StructMD} = Union{FMU2, FMU2Component, fmi2ModelDescription, FMU3, FMU3Component, fmi3ModelDescription}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

str::fmi2ModelDescription: Struct wich provides the static information of ModelVariables.

str::FMU3: Mutable struct representing an FMU in the FMI 3.0 Standard.

str::FMU3Component: Mutable struct represents a pointer to an FMU specific data structure that contains the information needed. Also in FMI 3.0 Standard.

str::fmi3ModelDescription: Struct witch provides the static information of ModelVariables.

See also FMU2, FMU2Component, fmi2ModelDescription, FMU3, FMU3Component, fmi3ModelDescription.

FMI.fmiGetGUID — Function

fmiGetGUID(str::fmi2StructMD)

Returns the tag 'guid' from the model description.

Arguments

str::fmi2StructMD: Representative for an FMU in the FMI 2.0.2 Standard. More detailed: fmi2StructMD = Union{FMU2, FMU2Component, fmi2ModelDescription}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

str::fmi2ModelDescription: Struct wich provides the static information of ModelVariables.

See also fmi2GetGUID, fmi2StructMD, FMU2, FMU2Component, fmi2ModelDescription.

FMI.fmiGetGenerationTool — Function

fmiGetGenerationTool(str::fmi2StructMD)

Returns the tag 'generationtool' from the model description.

Arguments

str::fmi2StructMD: Representative for an FMU in the FMI 2.0.2 Standard.

More detailed: fmi2StructMD = Union{FMU2, FMU2Component, fmi2ModelDescription}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

str::fmi2ModelDescription: Struct wich provides the static information of ModelVariables.

Returns

str.generationtool: The function fmi2GetGenerationTool returns the tag 'generationtool' from the struct, representing a FMU (str).

See also fmi2GetGenerationTool, fmi2StructMD, FMU2, FMU2Component, fmi2ModelDescription.

FMI.fmiGetGenerationDateAndTime — Function

fmiGetGenerationDateAndTime(str::fmi2StructMD)

Returns the tag 'generationdateandtime' from the model description.

Arguments

str::fmi2StructMD: Representative for an FMU in the FMI 2.0.2 Standard.

More detailed: fmi2StructMD = Union{FMU2, FMU2Component, fmi2ModelDescription}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

str::fmi2ModelDescription: Struct witch provides the static information of ModelVariables.

Returns

str.generationDateAndTime: The function fmi2GetGenerationDateAndTime returns the tag 'generationDateAndTime' from the struct, representing a FMU (str).

See also fmi2GetGenerationDateAndTime, fmi2StructMD, FMU2, FMU2Component, fmi2ModelDescription.

FMI.fmiGetVariableNamingConvention — Function

fmiGetVariableNamingConvention(str::fmi2StructMD)

Returns the tag 'varaiblenamingconvention' from the model description.

Arguments

str::fmi2StructMD: Representative for an FMU in the FMI 2.0.2 Standard.

More detailed: fmi2StructMD = Union{FMU2, FMU2Component, fmi2ModelDescription}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

str::fmi2ModelDescription: Struct witch provides the static information of ModelVariables.

Returns

str.variableNamingConvention: The function fmi2GetVariableNamingConvention returns the tag 'variableNamingConvention' from the struct, representing a FMU (str).

See also fmi2GetVariableNamingConvention, fmi2StructMD, FMU2, FMU2Component, fmi2ModelDescription.

FMI.fmiGetNumberOfEventIndicators — Function

fmiGetNumberOfEventIndicators(str::fmi2StructMD)

Returns the tag 'numberOfEventIndicators' from the model description.

Arguments

str::fmi2StructMD: Representative for an FMU in the FMI 2.0.2 Standard.

More detailed: fmi2StructMD = Union{FMU2, FMU2Component, fmi2ModelDescription}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

str::fmi2ModelDescription: Struct witch provides the static information of ModelVariables.

Returns

str.numberOfEventIndicators: The function fmi2GetNumberOfEventIndicators returns the tag 'numberOfEventIndicators' from the struct, representing a FMU (str).

See also fmi2GetNumberOfEventIndicators, fmi2StructMD, FMU2, FMU2Component, fmi2ModelDescription.

FMI.fmiCanGetSetState — Function

fmiCanGetSetState(str::fmi2StructMD)

Returns true, if the FMU supports the getting/setting of states

Arguments

str::fmi2StructMD: Representative for an FMU in the FMI 2.0.2 Standard.

More detailed: fmi2StructMD = Union{FMU2, FMU2Component, fmi2ModelDescription}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

str::fmi2ModelDescription: Struct witch provides the static information of ModelVariables.

Returns

::Bool: The function fmi2CanGetSetState returns True, if the FMU supports the getting/setting of states.

See also fmi2CanGetSetState, fmi2StructMD, FMU2, FMU2Component, fmi2ModelDescription.

FMI.fmiCanSerializeFMUstate — Function

fmiCanSerializeFMUstate(str::fmi2StructMD)

Returns true, if the FMU state can be serialized

Arguments

str::fmi2StructMD: Representative for an FMU in the FMI 2.0.2 Standard.

More detailed: fmi2StructMD = Union{FMU2, FMU2Component, fmi2ModelDescription}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

str::fmi2ModelDescription: Struct wich provides the static information of ModelVariables.

Returns

::Bool: The function fmi2CanSerializeFMUstate returns True, if the FMU state can be serialized.

See also fmi2CanSerializeFMUstate, fmi2StructMD, FMU2, FMU2Component, fmi2ModelDescription.

FMI.fmiProvidesDirectionalDerivative — Function

fmiProvidesDirectionalDerivative(str::fmi2StructMD)

Returns true, if the FMU provides directional derivatives

Arguments

str::fmi2StructMD: Representative for an FMU in the FMI 2.0.2 Standard.

More detailed: fmi2StructMD = Union{FMU2, FMU2Component, fmi2ModelDescription}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

str::fmi2ModelDescription: Struct witch provides the static information of ModelVariables.

Returns

::Bool: The function fmi2ProvidesDirectionalDerivative returns True, if the FMU provides directional derivatives.

See also fmi2ProvidesDirectionalDerivative, fmi2StructMD, FMU2, FMU2Component, fmi2ModelDescription.

FMI.fmiIsCoSimulation — Function

fmiIsCoSimulation(str::fmi2StructMD)

Returns true, if the FMU supports co simulation

Arguments

str::fmi2StructMD: Representative for an FMU in the FMI 2.0.2 Standard.

More detailed: fmi2StructMD = Union{FMU2, FMU2Component, fmi2ModelDescription}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

str::fmi2ModelDescription: Struct witch provides the static information of ModelVariables.

Returns

::Bool: The function fmi2IsCoSimulation returns True, if the FMU supports co simulation

See also fmi2IsCoSimulation, fmi2StructMD, FMU2, FMU2Component, fmi2ModelDescription.

FMI.fmiIsModelExchange — Function

fmiIsModelExchange(str::fmi2StructMD)

Returns true, if the FMU supports model exchange

Arguments

str::fmi2StructMD: Representative for an FMU in the FMI 2.0.2 Standard.

More detailed: fmi2StructMD = Union{FMU2, FMU2Component, fmi2ModelDescription}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

str::fmi2ModelDescription: Struct witch provides the static information of ModelVariables.

Returns

::Bool: The function fmi2IsCoSimulation returns True, if the FMU supports model exchange.

See also fmi2IsModelExchange, fmi2StructMD, FMU2, FMU2Component, fmi2ModelDescription.

FMI.fmiInfo — Function

fmiInfo(str::fmi2Struct)

Prints FMU-specific information into the REPL.

Arguments

str::fmi2Struct: Representative for an FMU in the FMI 2.0.2 Standard.

More detailed: fmi2Struct = Union{FMU2, FMU2Component}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

See also fmi2Info, fmi2Struct, FMU2, FMU2Component.

Get/Set variable values

FMI.fmiGet — Function

fmiGet(fmu::FMU2, args...; kwargs...)

Wrapper for fmiGet() in FMIImport/FMI2_ext.jl

fmiGet(str::fmi2Struct, args...; kwargs...)

FMI.fmiGet! — Function

fmiGet!(fmu::FMU2, args...; kwargs...)

Wrapper for fmiGet!() in FMIImport/FMI2_ext.jl

ToDo

FMI.fmiSet — Function

fmiSet(fmu::FMU2, args...; kwargs...)

Wrapper for fmiSet() in FMIImport/FMI2_ext.jl

ToDo

Simulate FMU

FMI.fmiSimulate — Function

fmiSimulate(str::fmi2Struct, t_start::Union{Real, Nothing} = nothing, t_stop::Union{Real, Nothing} = nothing;
            tolerance::Union{Real, Nothing} = nothing,
            dt::Union{Real, Nothing} = nothing,
            solver = nothing,
            customFx = nothing,
            recordValues::fmi2ValueReferenceFormat = nothing,
            saveat = [],
            setup::Bool = true,
            reset::Union{Bool, Nothing} = nothing, # nothing = auto
            inputValueReferences::fmi2ValueReferenceFormat = nothing,
            inputFunction = nothing,
            parameters::Union{Dict{<:Any, <:Any}, Nothing} = nothing,
            dtmax::Union{Real, Nothing} = nothing,
            kwargs...)

Starts a simulation of the FMU instance for the matching FMU type, if both types are available, CS is preferred.

Arguments

str::fmi2Struct: Representative for an FMU in the FMI 2.0.2 Standard.

More detailed: fmi2Struct = Union{FMU2, FMU2Component}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

t_start::Union{Real, Nothing} = nothing: Set the start time to a value of type Real or the default value from the model description is used.

t_stop::Union{Real, Nothing} = nothing: Set the end time to a value of type Real or the default value from the model description is used.

Keywords

tolerance::Union{Real, Nothing} = nothing: Real number to set the tolerance for any OED-solver

dt::Union{Real, Nothing} = nothing: Real number to set the step size of the OED-solver. Defaults to an automatic choice if the method is adaptive. More Info: DifferentialEquations.jl Documentation

solver = nothing: Any Julia-supported OED-solver (default is Tsit5). More Info: DifferentialEquations.jl Documentation

customFx = nothing: [deperecated] Ability to give a custom state derivative function ẋ=f(x,t)

recordValues::fmi2ValueReferenceFormat = nothing: AbstractArray of variables (strings or variableIdentifiers) to record. Results are returned as DiffEqCallbacks.SavedValues

saveat = []: Time points to save values at (interpolated). More Info: DifferentialEquations.jl Documentation

setup::Bool = true: Boolean, if FMU should be setup (default: setup=true)

reset::Union{Bool, Nothing} = nothing: Boolean, if FMU should be reset before simulation (default: reset:=auto)

inputValueReferences::fmi2ValueReferenceFormat = nothing: AbstractArray of input variables (strings or variableIdentifiers) to set at every simulation step

inputFunction = nothing: Function to retrieve the values to set the inputs to

parameters::Union{Dict{<:Any, <:Any}, Nothing} = nothing: Dictionary of parameter variables (strings or variableIdentifiers) and values (Real, Integer, Boolean, String) to set parameters during initialization

dtmax::Union{Real, Nothing} = nothing: Real number for setting maximum dt for adaptive timestepping for the ODE solver. The default values are package dependent. More Info: DifferentialEquations.jl Documentation

kwargs...: Further parameters of already defined functions solve(args..., kwargs...) from the library DifferentialEquations.jl

Returns

success::Bool for CS-FMUs

ODESolution for ME-FMUs

if keyword recordValues is set, a tuple of type (success::Bool, DiffEqCallbacks.SavedValues) for CS-FMUs

if keyword recordValues is set, a tuple of type (ODESolution, DiffEqCallbacks.SavedValues) for ME-FMUs

See also fmi2Simulate, fmi2SimulateME, fmi2SimulateCS, fmi2Struct, FMU2, FMU2Component.

FMI.fmiSimulateCS — Function

fmiSimulateCS(str::fmi2Struct, t_start::Union{Real, Nothing} = nothing, t_stop::Union{Real, Nothing} = nothing;
            tolerance::Union{Real, Nothing} = nothing,
            dt::Union{Real, Nothing} = nothing,
            solver = nothing,
            customFx = nothing,
            recordValues::fmi2ValueReferenceFormat = nothing,
            saveat = [],
            setup::Bool = true,
            reset::Union{Bool, Nothing} = nothing, # nothing = auto
            inputValueReferences::fmi2ValueReferenceFormat = nothing,
            inputFunction = nothing,
            parameters::Union{Dict{<:Any, <:Any}, Nothing} = nothing,
            dtmax::Union{Real, Nothing} = nothing,
            kwargs...)

Starts a simulation of the Co-Simulation FMU instance.

Arguments

str::fmi2Struct: Representative for an FMU in the FMI 2.0.2 Standard.

More detailed: fmi2Struct = Union{FMU2, FMU2Component}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

t_start::Union{Real, Nothing} = nothing: Set the start time to a value of type Real or the default value from the model description is used.

t_stop::Union{Real, Nothing} = nothing: Set the end time to a value of type Real or the default value from the model description is used.

Keywords

tolerance::Union{Real, Nothing} = nothing: Real number to set the tolerance for any OED-solver

dt::Union{Real, Nothing} = nothing: Real number to set the step size of the OED-solver. Defaults to an automatic choice if the method is adaptive. More Info: DifferentialEquations.jl Documentation

solver = nothing: Any Julia-supported OED-solver (default is Tsit5). More Info: DifferentialEquations.jl Documentation

customFx = nothing: [deperecated] Ability to give a custom state derivative function ẋ=f(x,t)

recordValues::fmi2ValueReferenceFormat = nothing: AbstractArray of variables (strings or variableIdentifiers) to record. Results are returned as DiffEqCallbacks.SavedValues

saveat = []: Time points to save values at (interpolated). More Info: DifferentialEquations.jl Documentation

setup::Bool = true: Boolean, if FMU should be setup (default: setup=true)

reset::Union{Bool, Nothing} = nothing: Boolean, if FMU should be reset before simulation (default: reset:=auto)

inputValueReferences::fmi2ValueReferenceFormat = nothing: AbstractArray of input variables (strings or variableIdentifiers) to set at every simulation step

inputFunction = nothing: Function to retrieve the values to set the inputs to

parameters::Union{Dict{<:Any, <:Any}, Nothing} = nothing: Dictionary of parameter variables (strings or variableIdentifiers) and values (Real, Integer, Boolean, String) to set parameters during initialization

dtmax::Union{Real, Nothing} = nothing: Real number for setting maximum dt for adaptive timestepping for the ODE solver. The default values are package dependent. More Info: DifferentialEquations.jl Documentation

kwargs...: Further parameters of already defined functions solve(args..., kwargs...) from the library DifferentialEquations.jl

Returns

If keyword recordValues is not set, a boolean success is returned (simulation success).

If keyword recordValues is set, a tuple of type (true, DiffEqCallbacks.SavedValues) or (false, nothing).

See also fmi2SimulateCS, fmi2Simulate, fmi2SimulateME, fmi2Struct, FMU2, FMU2Component.

FMI.fmiSimulateME — Function

fmiSimulateME(str::fmi2Struct, t_start::Union{Real, Nothing} = nothing, t_stop::Union{Real, Nothing} = nothing;
            tolerance::Union{Real, Nothing} = nothing,
            dt::Union{Real, Nothing} = nothing,
            solver = nothing,
            customFx = nothing,
            recordValues::fmi2ValueReferenceFormat = nothing,
            saveat = [],
            setup::Bool = true,
            reset::Union{Bool, Nothing} = nothing, # nothing = auto
            inputValueReferences::fmi2ValueReferenceFormat = nothing,
            inputFunction = nothing,
            parameters::Union{Dict{<:Any, <:Any}, Nothing} = nothing,
            dtmax::Union{Real, Nothing} = nothing,
            kwargs...)

Simulates a FMU instance for the given simulation time interval.

Arguments

str::fmi2Struct: Representative for an FMU in the FMI 2.0.2 Standard.

More detailed: fmi2Struct = Union{FMU2, FMU2Component}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

t_start::Union{Real, Nothing} = nothing: Set the start time to a value of type Real or the default value from the model description is used.

t_stop::Union{Real, Nothing} = nothing: Set the end time to a value of type Real or the default value from the model description is used.

Keywords

tolerance::Union{Real, Nothing} = nothing: Real number to set the tolerance for any OED-solver

dt::Union{Real, Nothing} = nothing: Real number to set the step size of the OED-solver. Defaults to an automatic choice if the method is adaptive. More Info: DifferentialEquations.jl Documentation

solver = nothing: Any Julia-supported OED-solver (default is Tsit5). More Info: DifferentialEquations.jl Documentation

customFx = nothing: [deperecated] Ability to give a custom state derivative function ẋ=f(x,t)

recordValues::fmi2ValueReferenceFormat = nothing: AbstractArray of variables (strings or variableIdentifiers) to record. Results are returned as DiffEqCallbacks.SavedValues

saveat = []: Time points to save values at (interpolated). More Info: DifferentialEquations.jl Documentation

setup::Bool = true: Boolean, if FMU should be setup (default: setup=true)

reset::Union{Bool, Nothing} = nothing: Boolean, if FMU should be reset before simulation (default: reset:=auto)

inputValueReferences::fmi2ValueReferenceFormat = nothing: AbstractArray of input variables (strings or variableIdentifiers) to set at every simulation step

inputFunction = nothing: Function to retrieve the values to set the inputs to

parameters::Union{Dict{<:Any, <:Any}, Nothing} = nothing: Dictionary of parameter variables (strings or variableIdentifiers) and values (Real, Integer, Boolean, String) to set parameters during initialization

dtmax::Union{Real, Nothing} = nothing: Real number for setting maximum dt for adaptive timestepping for the ODE solver. The default values are package dependent. More Info: DifferentialEquations.jl Documentation

kwargs...: Further parameters of already defined functions solve(args..., kwargs...) from the library DifferentialEquations.jl

Returns

If keyword recordValues is not set, a struct of type ODESolution.

If keyword recordValues is set, a tuple of type (ODESolution, DiffEqCallbacks.SavedValues).

See also fmi2SimulateME fmi2SimulateCS, fmi2Simulate, fmi2Struct, FMU2, FMU2Component.

Visualize simulation results

Missing docstring.

Missing docstring for fmiPlot. Check Documenter's build log for details.