FMU Standard 2 functions

Simulation results

FMI.fmi2SimulationResultGetValuesAtIndex — Function

Collects all data points for variable save index ´i´ inside a fmi2SimulationResult ´sd´.

FMI.fmi2SimulationResultGetTime — Function

Collects all time data points inside a fmi2SimulationResult ´sd´.

FMI.fmi2SimulationResultGetValues — Function

Collects all data points for variable with value reference ´tvr´ inside a fmi2SimulationResult ´sd´.

Collects all data points for variable with value name ´s´ inside a fmi2SimulationResult ´sd´.

Parsing variable names to ValueReferences

FMI.fmi2String2ValueReference — Function

Returns an array of ValueReferences coresponding to the variable names.

Returns the ValueReference coresponding to the variable name.

FMI.fmi2ValueReference2String — Function

Returns an array of variable names matching a fmi2ValueReference.

Opening and closing FMUs

FMI.fmi2Load — Function

Sets the properties of the fmu by reading the modelDescription.xml. Retrieves all the pointers of binary functions.

Returns the instance of the FMU struct.

Via optional argument unpackPath, a path to unpack the FMU can be specified (default: system temporary directory).

FMI.fmi2Unload — Function

Unload a FMU.

Free the allocated memory, close the binaries and remove temporary zip and unziped FMU model description.

Reading the model description

FMI.fmi2GetModelName — Function

Returns the tag 'modelName' from the model description.

FMI.fmi2GetGUID — Function

Returns the tag 'guid' from the model description.

FMI.fmi2GetGenerationTool — Function

Returns the tag 'generationtool' from the model description.

FMI.fmi2GetGenerationDateAndTime — Function

Returns the tag 'generationdateandtime' from the model description.

FMI.fmi2GetVariableNamingConvention — Function

Returns the tag 'varaiblenamingconvention' from the model description.

FMI.fmi2GetNumberOfEventIndicators — Function

Returns the tag 'numberOfEventIndicators' from the model description.

FMI.fmi2CanGetSetState — Function

Returns true, if the FMU supports the getting/setting of states

FMI.fmi2CanSerializeFMUstate — Function

Returns true, if the FMU state can be serialized

FMI.fmi2ProvidesDirectionalDerivative — Function

Returns true, if the FMU provides directional derivatives

FMI.fmi2IsCoSimulation — Function

Returns true, if the FMU supports co simulation

FMI.fmi2IsModelExchange — Function

Returns true, if the FMU supports model exchange

FMI.fmi2Info — Function

Prints FMU related information.

Simulate FMU

FMI.fmi2Simulate — Function

Starts a simulation of the fmu instance for the matching fmu type. If both types are available, CS is preferred over ME.

FMI.fmi2SimulateCS — Method

Starts a simulation of the Co-Simulation FMU instance.

Returns a tuple of (success::Bool, DiffEqCallbacks.SavedValues) with success = true or false. If keyword recordValues is not set, a tuple of type (success::Bool, nothing) is returned for consitency.

ToDo: Improove Documentation.

FMI.fmi2SimulateME — Function

Source: FMISpec2.0.2[p.90 ff]: 3.2.4 Pseudocode Example

Simulates a FMU instance for the given simulation time interval. State- and Time-Events are handled correctly.

Returns a tuple of type (ODESolution, DiffEqCallbacks.SavedValues). If keyword recordValues is not set, a tuple of type (ODESolution, nothing) is returned for consitency.