Library functions

Creation, Destruction FMU

Missing docstring.

Missing docstring for fmi2Instantiate. Check Documenter's build log for details.

FMI.fmi2FreeInstance! — Method

Source: FMISpec2.0.2[p.22]: 2.1.5 Creation, Destruction and Logging of FMU Instances

Disposes the given instance, unloads the loaded model, and frees all the allocated memory and other resources that have been allocated by the functions of the FMU interface. If a null pointer is provided for “c”, the function call is ignored (does not have an effect).

Removes the component from the FMUs component list.

Platform and Version number

FMI.fmi2GetTypesPlatform — Method

Source: FMISpec2.0.2[p.22]: 2.1.4 Inquire Platform and Version Number of Header Files

Returns the string to uniquely identify the “fmi2TypesPlatform.h” header file used for compilation of the functions of the FMU. The standard header file, as documented in this specification, has fmi2TypesPlatform set to “default” (so this function usually returns “default”).

FMI.fmi2GetVersion — Method

Source: FMISpec2.0.2[p.22]: 2.1.4 Inquire Platform and Version Number of Header Files

Returns the version of the “fmi2Functions.h” header file which was used to compile the functions of the FMU. The function returns “fmiVersion” which is defined in this header file. The standard header file as documented in this specification has version “2.0”

FMI.fmi2SetDebugLogging — Method

Source: FMISpec2.0.2[p.22]: 2.1.5 Creation, Destruction and Logging of FMU Instances

The function controls debug logging that is output via the logger function callback. If loggingOn = fmi2True, debug logging is enabled, otherwise it is switched off.

Initialization, Termination and Destruction

FMI.fmi2SetupExperiment — Method

Source: FMISpec2.0.2[p.23]: 2.1.6 Initialization, Termination, and Resetting an FMU

Informs the FMU to setup the experiment. This function must be called after fmi2Instantiate and before fmi2EnterInitializationMode is called.The function controls debug logging that is output via the logger function callback. If loggingOn = fmi2True, debug logging is enabled, otherwise it is switched off.

FMI.fmi2EnterInitializationMode — Method

Source: FMISpec2.0.2[p.23]: 2.1.6 Initialization, Termination, and Resetting an FMU

Informs the FMU to enter Initialization Mode. Before calling this function, all variables with attribute <ScalarVariable initial = "exact" or "approx"> can be set with the “fmi2SetXXX” functions (the ScalarVariable attributes are defined in the Model Description File, see section 2.2.7). Setting other variables is not allowed. Furthermore, fmi2SetupExperiment must be called at least once before calling fmi2EnterInitializationMode, in order that startTime is defined.

FMI.fmi2ExitInitializationMode — Method

Source: FMISpec2.0.2[p.23]: 2.1.6 Initialization, Termination, and Resetting an FMU

Informs the FMU to exit Initialization Mode.

FMI.fmi2Terminate — Method

Source: FMISpec2.0.2[p.24]: 2.1.6 Initialization, Termination, and Resetting an FMU

Informs the FMU that the simulation run is terminated.

FMI.fmi2Reset — Method

Source: FMISpec2.0.2[p.24]: 2.1.6 Initialization, Termination, and Resetting an FMU

Is called by the environment to reset the FMU after a simulation run. The FMU goes into the same state as if fmi2Instantiate would have been called.

Get/Set variable values

FMI.fmi2GetReal — Method

TODO: FMI specification reference.

Get the values of an array of fmi2Real variables.

For more information call ?fmi2GetReal!

Missing docstring.

Missing docstring for fmi2GetReal!(::FMI.fmi2Component, ::Array{FMI.fmi2ValueReference}, ::Csize_t, ::Array{FMI.fmi2Real}). Check Documenter's build log for details.

FMI.fmi2SetReal — Method

Source: FMISpec2.0.2[p.24]: 2.1.7 Getting and Setting Variable Values

Functions to get and set values of variables idetified by their valueReference

FMI.fmi2GetInteger — Method

TODO: FMI specification reference.

Get the values of an array of fmi2Integer variables.

For more information call ?fmi2GetInteger!

Missing docstring.

Missing docstring for fmi2GetInteger!(::FMI.fmi2Component, ::Array{FMI.fmi2ValueReference}, ::Csize_t, ::Array{FMI.fmi2Integer}). Check Documenter's build log for details.

FMI.fmi2SetInteger — Method

Source: FMISpec2.0.2[p.24]: 2.1.7 Getting and Setting Variable Values

Functions to get and set values of variables idetified by their valueReference

FMI.fmi2GetBoolean — Method

TODO: FMI specification reference.

Get the values of an array of fmi2Boolean variables.

For more information call ?fmi2GetBoolean!

Missing docstring.

Missing docstring for fmi2GetBoolean!(::FMI.fmi2Component, ::Array{FMI.fmi2ValueReference}, ::Csize_t, ::Array{FMI.fmi2Boolean}). Check Documenter's build log for details.

FMI.fmi2SetBoolean — Method

Source: FMISpec2.0.2[p.24]: 2.1.7 Getting and Setting Variable Values

Functions to get and set values of variables idetified by their valueReference

FMI.fmi2GetString — Method

TODO: FMI specification reference.

Get the values of an array of fmi2String variables.

For more information call ?fmi2GetString!

Missing docstring.

Missing docstring for fmi2GetString!(::FMI.fmi2Component, ::Array{FMI.fmi2ValueReference}, ::Csize_t, ::Vector{Ptr{Cchar}}). Check Documenter's build log for details.

FMI.fmi2SetString — Method

Source: FMISpec2.0.2[p.24]: 2.1.7 Getting and Setting Variable Values

Functions to get and set values of variables idetified by their valueReference

FMU state Functions

FMI.fmi2GetFMUstate — Method

Source: FMISpec2.0.2[p.26]: 2.1.8 Getting and Setting the Complete FMU State

fmi2GetFMUstate makes a copy of the internal FMU state and returns a pointer to this copy

FMI.fmi2SetFMUstate — Method

Source: FMISpec2.0.2[p.26]: 2.1.8 Getting and Setting the Complete FMU State

fmi2SetFMUstate copies the content of the previously copied FMUstate back and uses it as actual new FMU state.

FMI.fmi2FreeFMUstate — Method

Source: FMISpec2.0.2[p.26]: 2.1.8 Getting and Setting the Complete FMU State

fmi2FreeFMUstate frees all memory and other resources allocated with the fmi2GetFMUstate call for this FMUstate.

FMI.fmi2SerializedFMUstateSize — Method

Source: FMISpec2.0.2[p.26]: 2.1.8 Getting and Setting the Complete FMU State

fmi2SerializedFMUstateSize returns the size of the byte vector, in order that FMUstate can be stored in it.

FMI.fmi2SerializeFMUstate — Method

Source: FMISpec2.0.2[p.26]: 2.1.8 Getting and Setting the Complete FMU State

fmi2SerializeFMUstate serializes the data which is referenced by pointer FMUstate and copies this data in to the byte vector serializedState of length size

FMI.fmi2DeSerializeFMUstate — Method

Source: FMISpec2.0.2[p.26]: 2.1.8 Getting and Setting the Complete FMU State

fmi2DeSerializeFMUstate deserializes the byte vector serializedState of length size, constructs a copy of the FMU state and returns FMUstate, the pointer to this copy.

Partial Derivatives

FMI.fmi2GetDirectionalDerivative! — Method

Source: FMISpec2.0.2[p.26]: 2.1.9 Getting Partial Derivatives

This function computes the directional derivatives of an FMU.

FMI.fmi2GetDirectionalDerivative — Function

TODO: FMI specification reference.

Computes directional derivatives.

For more information call ?fmi2GetDirectionalDerivatives

CoSimulation specific Functions

FMI.fmi2SetRealInputDerivatives — Method

Source: FMISpec2.0.2[p.104]: 4.2.1 Transfer of Input / Output Values and Parameters

Sets the n-th time derivative of real input variables. vr defines the value references of the variables the array order specifies the corresponding order of derivation of the variables

FMI.fmi2GetRealOutputDerivatives — Method

Source: FMISpec2.0.2[p.104]: 4.2.1 Transfer of Input / Output Values and Parameters

Retrieves the n-th derivative of output values. vr defines the value references of the variables the array order specifies the corresponding order of derivation of the variables

FMI.fmi2DoStep — Method

Source: FMISpec2.0.2[p.104]: 4.2.2 Computation

The computation of a time step is started.

FMI.fmi2CancelStep — Method

Source: FMISpec2.0.2[p.105]: 4.2.2 Computation

Can be called if fmi2DoStep returned fmi2Pending in order to stop the current asynchronous execution.

FMI.fmi2GetStatus — Method

Source: FMISpec2.0.2[p.106]: 4.2.3 Retrieving Status Information from the Slave

Informs the master about the actual status of the simulation run. Which status information is to be returned is specified by the argument fmi2StatusKind.

FMI.fmi2GetRealStatus — Method

Source: FMISpec2.0.2[p.106]: 4.2.3 Retrieving Status Information from the Slave

Informs the master about the actual status of the simulation run. Which status information is to be returned is specified by the argument fmi2StatusKind.

FMI.fmi2GetIntegerStatus — Method

Source: FMISpec2.0.2[p.106]: 4.2.3 Retrieving Status Information from the Slave

Informs the master about the actual status of the simulation run. Which status information is to be returned is specified by the argument fmi2StatusKind.

FMI.fmi2GetBooleanStatus — Method

Source: FMISpec2.0.2[p.106]: 4.2.3 Retrieving Status Information from the Slave

Informs the master about the actual status of the simulation run. Which status information is to be returned is specified by the argument fmi2StatusKind.

FMI.fmi2GetStringStatus — Method

Source: FMISpec2.0.2[p.106]: 4.2.3 Retrieving Status Information from the Slave

Informs the master about the actual status of the simulation run. Which status information is to be returned is specified by the argument fmi2StatusKind.

ModelExchange specific Functions

FMI.fmi2SetTime — Method

Source: FMISpec2.0.2[p.83]: 3.2.1 Providing Independent Variables and Re-initialization of Caching

Set a new time instant and re-initialize caching of variables that depend on time, provided the newly provided time value is different to the previously set time value (variables that depend solely on constants or parameters need not to be newly computed in the sequel, but the previously computed values can be reused).

FMI.fmi2SetContinuousStates — Method

Source: FMISpec2.0.2[p.83]: 3.2.1 Providing Independent Variables and Re-initialization of Caching

Set a new (continuous) state vector and re-initialize caching of variables that depend on the states. Argument nx is the length of vector x and is provided for checking purposes

FMI.fmi2EnterEventMode — Method

Source: FMISpec2.0.2[p.84]: 3.2.2 Evaluation of Model Equations

The model enters Event Mode from the Continuous-Time Mode and discrete-time equations may become active (and relations are not “frozen”).

FMI.fmi2NewDiscreteStates — Method

Source: FMISpec2.0.2[p.84]: 3.2.2 Evaluation of Model Equations

The FMU is in Event Mode and the super dense time is incremented by this call.

FMI.fmi2EnterContinuousTimeMode — Method

Source: FMISpec2.0.2[p.85]: 3.2.2 Evaluation of Model Equations

The model enters Continuous-Time Mode and all discrete-time equations become inactive and all relations are “frozen”. This function has to be called when changing from Event Mode (after the global event iteration in Event Mode over all involved FMUs and other models has converged) into Continuous-Time Mode.

Missing docstring.

Missing docstring for fmi2CompletedIntegratorStep!(::FMI.fmi2Component, ::FMI.fmi2Boolean, ::FMI.fmi2Boolean, ::FMI.fmi2Boolean). Check Documenter's build log for details.

FMI.fmi2GetDerivatives — Method

Source: FMISpec2.0.2[p.86]: 3.2.2 Evaluation of Model Equations

Compute state derivatives at the current time instant and for the current states.

FMI.fmi2GetEventIndicators — Method

Source: FMISpec2.0.2[p.86]: 3.2.2 Evaluation of Model Equations

Compute event indicators at the current time instant and for the current states.

FMI.fmi2GetContinuousStates — Method

Source: FMISpec2.0.2[p.86]: 3.2.2 Evaluation of Model Equations

Return the new (continuous) state vector x.

FMI.fmi2GetNominalsOfContinuousStates — Method

Source: FMISpec2.0.2[p.86]: 3.2.2 Evaluation of Model Equations

Return the nominal values of the continuous states.